ModelPredictiveControl.jl Documentation

ModelPredictiveControl.LinModel — Type

LinModel(sys::StateSpace, Ts::Real; kwargs...)

Construct a LinModel from state-state model sys.

If sys is continuous, it is dicretized using c2d and :zoh for manipulated inputs, and :tustin, for measured disturbances.

Arguments

sys::StateSpace: state-space model incl. manipulated inputs and measured disturbances

Ts::Real: model sampling time in second

i_u::IntRangeOrVector = 1:size(sys,2): index of sys inputs that are manipulated

i_d::IntRangeOrVector = Int[]: index of sys inputs that are measured disturbances

u_op::Vector{<:Real} = Float64[]: manipulated input operating points

y_op::Vector{<:Real} = Float64[]: outputs operating points

d_op::Vector{<:Real} = Float64[]: measured disturbances operating points

Examples

julia> LinModel(tf(3, [10, 1]), 2, u_op=[50], y_op=[20])
Discrete-time linear model with a sample time Ts = 2.0 s and:
- 1 manipulated inputs u
- 1 states x
- 1 outputs y
- 0 measured disturbances d

LinModel(sys::TransferFunction, Ts::Real; kwargs...)

Convert to minimal realization state-space when sys is a transfer function.