Functions: SimModel Internals

Functions: SimModel Internals

Abstract Types

Model Construction

State-Space Functions

Init State

Abstract Types

ModelPredictiveControl.SimModelODE — Type

abstract type SimModelODE <: SimModel

Abstract subtype of SimModel for ordinary differential equations.

ModelPredictiveControl.SimModelDAE — Type

abstract type SimModelDAE <: SimModel

Abstract subtype of SimModel for differential and algebraic equations.

Model Construction

ModelPredictiveControl.init_defectmat_dae — Function

init_defectmat_dae(NT, ::OrthogonalCollocation, nx, na, Co, λo) -> Es, Ks, Aeq

Init the matrices for computing the defect of the next state.

Knowing that the decision vector $\mathbf{Z}$ contain $\mathbf{x̂_0}(k+1)$, $\mathbf{a_0}(k+0)$, $\mathbf{k̄}(k+0)$ and $\mathbf{ā}(k+0)$ vectors with an OrthogonalCollocation, this linear equation compute the defect of the states at time $k+1$:

\[\begin{aligned} \mathbf{s}(k+1) &= \mathbf{E_s Z + K_s x_0}(k) \\ &= \mathbf{E_s Z + F_s} \end{aligned}\]

It is forced to be $\mathbf{s}(k+1) = \mathbf{0}$ using the optimization equality constraints.

init_defectmat_dae(NT, ::CollocationMethod, nx, na, _ , _ ) -> Es, Ks, Aeq

No linear equality constraint for other CollocationMethods, return empty matrices.

State-Space Functions

ModelPredictiveControl.f! — Function

f!(x0next, _ , model::LinModel, x0, u0, d0, _ ) -> nothing

Evaluate x0next = A*x0 + Bu*u0 + Bd*d0 in-place when model is a LinModel.

f!(x0next, k̄, model::NonLinModel, x0, u0, d0, p)

Compute x0next using the DiffSolver in model.solver and model.f!.

The method mutates x0next and k̄ arguments in-place. The latter is used to store the intermediate stage values of the solver.

f!(x0next, _ , model::NonLinModelDAE, x0, u0, d0, _ ) -> nothing

Solve the optimization model.optim problem for NonLinModelDAE.

After solving, the next state $\mathbf{x_0}(k+1)$ will be stored in-place in the x0next argument. The next algebraic variable $\mathbf{a_0}(k+1)$ will be also stored at model.a0.

ModelPredictiveControl.h! — Function

h!(y0, model::LinModel, x0, d0, _ ) -> nothing

Evaluate y0 = C*x0 + Dd*d0 in-place when model is a LinModel.

h!(y0, model::NonLinModel, x0, d0, p)

Compute y0 by calling model.h! directly for NonLinModel.

h!(y0, model::NonLinModelDAE, x0, d0, p) -> nothing

Solve the algebraic equation to get a0 and call model.h! for NonLinModelDAE.

If model.iszero_Ha is true, the algebraic variable is not used in model.h! according to SparseConnectivityTracer.jl, the algebraic equation solving is thus skipped and model.h! is called directly.

Init State

ModelPredictiveControl.initstate_core! — Function

initstate_core!(model::LinModel, u0, d0)

Set model.x0 to u0 and d0 steady-state if model is a LinModel.

Following setop! notation, the method evaluates the equilibrium from:

\[ \mathbf{x_0} = \mathbf{(I - A)^{-1}(B_u u_0 + B_d d_0 + f_{op} - x_{op})}\]

with constant manipulated inputs $\mathbf{u_0 = u - u_{op}}$ and measured disturbances $\mathbf{d_0 = d - d_{op}}$. The Moore-Penrose pseudo-inverse computes $\mathbf{(I - A)^{-1}}$ to support integrating model (integrator states will be 0).

initstate_core!(model::NonLinModelDAE, u0, d0)

Warm-start model.Z and model.a0 at zero if model is a NonLinModelDAE.

The field model.a0 and model.Z respectively warm-start evaloutput and updatestate! solving. The method also set model.optim_u0 and model.optim_d0 at u0 and d0 values. The model.u0 field is used to solve the algebraic equation $\mathbf{q}$ in evaloutput method, but it should not impact the result in theory since model is strictly proper w.r.t. u0.

initstate_core!(::SimModel, u0, d0)

Do nothing at all by default.

OrthogonalCollocation( h::Int=0, no::Int=3; f_threads=false, h_threads=false, roots=:gaussradau )
Construct an orthogonal collocation TranscriptionMethod.

LinModel(sys::StateSpace[, Ts]; i_u=1:size(sys,2), i_d=Int[]) LinModel(sys::TransferFunction[, Ts]; i_u=1:size(sys,2), i_d=Int[]) LinModel(sys::DelayLtiSystem, Ts; i_u=1:size(sys,2), i_d=Int[]) LinModel{NT}(A, Bu, C, Bd, Dd, Ts) LinModel(model::NonLinModel; x=model.x0+model.xop, u=model.uop, d=model.dop)
Construct a linear model from state-space model sys with sampling time Ts in seconds.

NonLinModel{NT}(f::Function, h::Function, Ts, nu, nx, ny, nd=0; <keyword arguments>) NonLinModel{NT}(f!::Function, h!::Function, Ts, nu, nx, ny, nd=0; <keyword arguments>)
Construct a nonlinear ODE model from state-space functions f/f! and h/h!.

NonLinModelDAE{NT}(fq::Function, h::Function, Ts, nu, nx, na, ny, nd=0; <kw args>) NonLinModelDAE{NT}(fq!::Function, h!::Function, Ts, nu, nx, na, ny, nd=0; <kw args>)
Construct a nonlinear DAE model from state-space functions fq/fq! and h/h!.

abstract type SimModel end
Supertype of LinModel and NonLinModel types.

abstract type CollocationMethod
Abstract subtype of TranscriptionMethod for direct collocation methods.