Model predictive controller design

asddas

PredictiveController functions and types

ModelPredictiveControl.LinMPC — Type

LinMPC(model::LinModel; <keyword arguments>)

Construct a linear predictive controller LinMPC based on LinModel model.

The controller minimizes the following objective function at each discrete time $k$:

\[\min_{\mathbf{ΔU}, ϵ} \mathbf{(R̂_y - Ŷ)}' \mathbf{M}_{H_p} \mathbf{(R̂_y - Ŷ)} + \mathbf{(ΔU)}' \mathbf{N}_{H_c} \mathbf{(ΔU)} + \mathbf{(R̂_u - U)}' \mathbf{L}_{H_p} \mathbf{(R̂_u - U)} + Cϵ^2\]

in which :

$H_p$ : prediction horizon

$H_c$ : control horizon

$\mathbf{ΔU}$ : manipulated input increments over $H_c$

$\mathbf{Ŷ}$ : predicted outputs over $H_p$

$\mathbf{U}$ : manipulated inputs over $H_p$

$\mathbf{R̂_y}$ : predicted output setpoints over $H_p$

$\mathbf{R̂_u}$ : predicted manipulated input setpoints over $H_p$

$\mathbf{M}_{H_p} = \text{diag}\mathbf{(M,M,...,M)}$ : output setpoint tracking weights

$\mathbf{N}_{H_c} = \text{diag}\mathbf{(N,N,...,N)}$ : manipulated input increment weights

$\mathbf{L}_{H_p} = \text{diag}\mathbf{(L,L,...,L)}$ : manipulated input setpoint tracking weights

$C$ : slack variable weight

$ϵ$ : slack variable for constraint softening

The $\mathbf{ΔU}$ vector includes the manipulated input increments $\mathbf{Δu}(k+j) = \mathbf{u}(k+j) - \mathbf{u}(k+j-1)$ from $j=0$ to $H_c-1$, the $\mathbf{Ŷ}$ vector, the output predictions $\mathbf{ŷ(k+j)}$ from $j=1$ to $H_p$, and the $\mathbf{U}$ vector, the manipulated inputs $\mathbf{u}(k+j)$ from $j=0$ to $H_p-1$. The manipulated input setpoint predictions $\mathbf{R̂_u}$ are constant at $\mathbf{r_u}$.

This method uses the default state estimator, a SteadyKalmanFilter with default arguments.

Arguments

model::LinModel : model used for controller predictions and state estimations.

Hp=10+nk: prediction horizon $H_p$, nk is the number of delays in model.

Hc=2 : control horizon $H_c$.

Mwt=fill(1.0,model.ny) : main diagonal of $\mathbf{M}$ weight matrix (vector)

Nwt=fill(0.1,model.nu) : main diagonal of $\mathbf{N}$ weight matrix (vector)

Lwt=fill(0.0,model.nu) : main diagonal of $\mathbf{L}$ weight matrix (vector)

Cwt=1e5 : slack variable weight $C$ (scalar), use Cwt=Inf for hard constraints only

ru=model.uop: manipulated input setpoints $\mathbf{r_u}$ (vector)

Extended Help

Manipulated inputs setpoints $\mathbf{r_u}$ are not common but they can be interesting for over-actuated systems, when nu > ny (e.g. prioritize solutions with lower economical costs). The default Lwt value implies that this feature is disabled by default.

LinMPC(estim::StateEstimator; <keyword arguments>)

Use custom state estimator estim to construct LinMPC.

estim.model must be a LinModel. Else, a NonLinMPC is required.

ModelPredictiveControl.setconstraint! — Function

setconstraint!(mpc::PredictiveController; <keyword arguments>)

Set the constraint parameters of mpc predictive controller.

The predictive controllers support both soft and hard constraints, defined by:

\[\begin{alignat*}{3} \mathbf{u_{min} - c_{u_{min}}} ϵ &≤ \mathbf{u}(k+j) &&≤ \mathbf{u_{max} + c_{u_{max}}} ϵ &&\qquad j = 0, 1 ,..., H_c - 1 \\ \mathbf{Δu_{min} - c_{Δu_{min}}} ϵ &≤ \mathbf{Δu}(k+j) &&≤ \mathbf{Δu_{max} + c_{Δu_{max}}} ϵ &&\qquad j = 0, 1 ,..., H_c - 1 \\ \mathbf{ŷ_{min} - c_{ŷ_{min}}} ϵ &≤ \mathbf{ŷ}(k+j) &&≤ \mathbf{ŷ_{max} + c_{ŷ_{max}}} ϵ &&\qquad j = 1, 2 ,..., H_p \\ \end{alignat*}\]

and also $ϵ ≥ 0$. All the constraint parameters are vector. Use ±Inf values when there is no bound. The constraint softness parameters $\mathbf{c}$, also called equal concern for relaxation, are non-negative values that specify the softness of the associated bound. Use 0.0 values for hard constraints. The predicted output constraints $\mathbf{ŷ_{min}}$ and $\mathbf{ŷ_{max}}$ are soft by default.

Arguments

Info

The default constraints are mentioned here for clarity but omitting a keyword argument will not re-assign to its default value (defaults are set at construction only).

umin=fill(-Inf,nu) : manipulated input lower bounds $\mathbf{u_{min}}$

umax=fill(+Inf,nu) : manipulated input upper bounds $\mathbf{u_{max}}$

Δumin=fill(-Inf,nu) : manipulated input increment lower bounds $\mathbf{Δu_{min}}$

Δumax=fill(+Inf,nu) : manipulated input increment upper bounds $\mathbf{Δu_{max}}$

ŷmin=fill(-Inf,ny) : predicted output lower bounds $\mathbf{ŷ_{min}}$

ŷmax=fill(+Inf,ny) : predicted output upper bounds $\mathbf{ŷ_{max}}$

c_umin=fill(0.0,nu) : umin softness weights $\mathbf{c_{u_{min}}}$

c_umax=fill(0.0,nu) : umax softness weights $\mathbf{c_{u_{max}}}$

c_Δumin=fill(0.0,nu) : Δumin softness weights $\mathbf{c_{Δu_{min}}}$

c_Δumax=fill(0.0,nu) : Δumax softness weights $\mathbf{c_{Δu_{max}}}$

c_ŷmin=fill(1.0,ny) : ŷmin softness weights $\mathbf{c_{ŷ_{min}}}$

c_ŷmax=fill(1.0,ny) : ŷmax softness weights $\mathbf{c_{ŷ_{max}}}$

Advanced Topics

Internals

ModelPredictiveControl.init_deterpred — Function

init_deterpred(model::LinModel, Hp, Hc)

Construct deterministic prediction matrices for LinModel model.

The linear model predictions are evaluated by :

\[\begin{aligned} \mathbf{Ŷ} &= \mathbf{E ΔU} + \mathbf{G d}(k) + \mathbf{J D̂} + \mathbf{K_d x̂_d}(k) + \mathbf{P u}(k-1) + \mathbf{Ŷ_s} \\ &= \mathbf{E ΔU} + \mathbf{F} \end{aligned}\]

where predicted outputs $\mathbf{Ŷ}$, stochastic outputs $\mathbf{Ŷ_s}$, and disturbances $\mathbf{D̂}$ are from $k + 1$ to $k + H_p$. Input increments $\mathbf{ΔU}$ are from $k$ to $k + H_c - 1$. Deterministic state estimates $\mathbf{x̂_d}(k)$ are extracted from current estimates $\mathbf{x̂}_{k-1}(k)$. Operating points on $\mathbf{u}$, $\mathbf{d}$ and $\mathbf{y}$ are omitted in above equations.

Note

Stochastic predictions $\mathbf{Ŷ_s}$ are calculated separately (see init_stochpred) and added to $\mathbf{F}$ matrix to support internal model structure and reduce NonLinMPC computational costs.

Extended Help

Using the $\mathbf{A, B_u, C, B_d, D_d}$ matrices in model and the equation $\mathbf{W}_j = \mathbf{C} ( ∑_{i=0}^j \mathbf{A}^i ) \mathbf{B_u}$, the prediction matrices are computed by :

\[\begin{aligned} \mathbf{E} &= \begin{bmatrix} \mathbf{W}_{0} & \mathbf{0} & \cdots & \mathbf{0} \\ \mathbf{W}_{1} & \mathbf{0} & \cdots & \mathbf{0} \\ \vdots & \vdots & \ddots & \vdots \\ \mathbf{W}_{H_p-1} & \mathbf{W}_{H_p-2} & \cdots & \mathbf{W}_{H_p-H_c+1} \end{bmatrix} \\ \mathbf{G} &= \begin{bmatrix} \mathbf{C}\mathbf{A}^{0} \mathbf{B_d} \\ \mathbf{C}\mathbf{A}^{1} \mathbf{B_d} \\ \vdots \\ \mathbf{C}\mathbf{A}^{H_p-1} \mathbf{B_d} \end{bmatrix} \\ \mathbf{J} &= \begin{bmatrix} \mathbf{D_d} & \mathbf{0} & \cdots & \mathbf{0} \\ \mathbf{C}\mathbf{A}^{0} \mathbf{B_d} & \mathbf{D_d} & \cdots & \mathbf{0} \\ \vdots & \vdots & \ddots & \mathbf{0} \\ \mathbf{C}\mathbf{A}^{H_p-2} \mathbf{B_d} & \mathbf{C}\mathbf{A}^{H_p-3} \mathbf{B_d} & \cdots & \mathbf{D_d} \end{bmatrix} \\ \mathbf{K_d} &= \begin{bmatrix} \mathbf{C}\mathbf{A}^{0} \\ \mathbf{C}\mathbf{A}^{1} \\ \vdots \\ \mathbf{C}\mathbf{A}^{H_p-1} \end{bmatrix} \\ \mathbf{P} &= \begin{bmatrix} \mathbf{W}_0 \\ \mathbf{W}_1 \\ \vdots \\ \mathbf{W}_{H_p-1} \end{bmatrix} \end{aligned}\]

ModelPredictiveControl.init_ΔUtoU — Function

init_ΔUtoU(nu, Hp, Hc, C, c_Umin, c_Umax)

Init manipulated input increments to inputs conversion matrices.

The conversion from the input increments $\mathbf{ΔU}$ to manipulated inputs over $H_p$ and $H_c$ are calculated by:

\[\begin{aligned} \mathbf{U} = \mathbf{U}_{H_p} &= \mathbf{S}_{H_p} \mathbf{ΔU} + \mathbf{T}_{H_p} \mathbf{u}(k-1) \\ \mathbf{U}_{H_c} &= \mathbf{S}_{H_c} \mathbf{ΔU} + \mathbf{T}_{H_c} \mathbf{u}(k-1) \end{aligned}\]

ModelPredictiveControl.relaxU — Function

relaxU(C, c_Umin, c_Umax, S_Hp, S_Hc)

Augment manipulated inputs constraints with slack variable ϵ for softening.

Denoting the input increments augmented with the slack variable $\mathbf{ΔŨ} = [\begin{smallmatrix} \mathbf{ΔU} \\ ϵ \end{smallmatrix}]$, it returns the augmented conversion matrices $\mathbf{S̃}_{H_p}$ and $\mathbf{S̃}_{H_c}$, similar to the ones described at init_ΔUtoU. It also returns the $\mathbf{A}$ matrices for the inequality constraints:

\[\begin{bmatrix} \mathbf{A_{u_{min}}} \\ \mathbf{A_{u_{max}}} \end{bmatrix} \mathbf{ΔŨ} ≤ \begin{bmatrix} + \mathbf{T}_{H_c} \mathbf{u}(k-1) - \mathbf{U_{min}} \\ - \mathbf{T}_{H_c} \mathbf{u}(k-1) + \mathbf{U_{max}} \end{bmatrix}\]

ModelPredictiveControl.relaxΔU — Function

relaxΔU(C, c_ΔUmin, c_ΔUmax, ΔUmin, ΔUmax, N_Hc)

Augment input increments constraints with slack variable ϵ for softening.

Denoting the input increments augmented with the slack variable $\mathbf{ΔŨ} = [\begin{smallmatrix} \mathbf{ΔU} \\ ϵ \end{smallmatrix}]$, it returns the augmented constraints $\mathbf{ΔŨ_{min}}$ and $\mathbf{ΔŨ_{max}}$ over $H_c$, and the $\mathbf{A}$ matrices for the inequality constraints:

\[\begin{bmatrix} \mathbf{A_{Δu_{min}}} \\ \mathbf{A_{Δu_{max}}} \end{bmatrix} \mathbf{ΔŨ} ≤ \begin{bmatrix} TODO: \end{bmatrix}\]

ModelPredictiveControl.relaxŶ — Function

relaxŶ(C, c_Ŷmin, c_Ŷmax, E)

Augment linear output prediction constraints with slack variable ϵ for softening.

Denoting the input increments augmented with the slack variable $\mathbf{ΔŨ} = [\begin{smallmatrix} \mathbf{ΔU} \\ ϵ \end{smallmatrix}]$, it returns the $\mathbf{Ẽ}$ matrix that appears in the linear model prediction equation $\mathbf{Ŷ = Ẽ ΔŨ + F}$, and the $\mathbf{A}$ matrices for the inequality constraints:

\[\begin{bmatrix} \mathbf{A_{ŷ_{min}}} \\ \mathbf{A_{ŷ_{max}}} \end{bmatrix} \mathbf{ΔŨ} ≤ \begin{bmatrix} + \mathbf{F} - \mathbf{Ŷ_{min}} \\ - \mathbf{F} + \mathbf{Ŷ_{max}} \end{bmatrix}\]

ModelPredictiveControl.init_quadprog — Function

init_quadprog(E, S_Hp, M_Hp, N_Hc, L_Hp)

Init the quadratic programming optimization matrix Q.

The Q matrix appears in the quadratic general form :

\[ J = \min_{\mathbf{ΔU}} \frac{1}{2}\mathbf{(ΔU)'Q(ΔU)} + \mathbf{p'(ΔU)} + q \]

$\mathbf{Q}$ is constant if the model and weights are linear and time invariant (LTI). The vector $\mathbf{p}$ and scalar $q$ need recalculation each control period $k$. $q$ does not impact the minima position. It is thus useless at optimization but required to evaluate the minimal $J$ value.

ModelPredictiveControl.init_stochpred — Function

init_stochpred(estim::StateEstimator, Hp)

Init the stochastic prediction matrix Ks from estim estimator for predictive control.

$\mathbf{K_s}$ is the prediction matrix of the stochastic model (composed exclusively of integrators):

\[ \mathbf{Ŷ_s} = \mathbf{K_s x̂_s}(k)\]

The stochastic predictions $\mathbf{Ŷ_s}$ are the integrator outputs from $k+1$ to $k+H_p$. $\mathbf{x̂_s}(k)$ is extracted from current estimates $\mathbf{x̂}_{k-1}(k)$. The method also returns the matrix $\mathbf{P_s = 0}$, which is useless except for [InternalModel] estimators.

init_stochpred(estim::InternalModel, Hp)

Init the stochastic prediction matrices for InternalModel.

Ks and Ps matrices are defined as:

\[ \mathbf{Ŷ_s} = \mathbf{K_s x̂_s}(k) + \mathbf{P_s ŷ_s}(k)\]

with $\mathbf{Ŷ_s}$ as stochastic predictions from $k+1$ to $k+H_p$, current stochastic states $\mathbf{x̂_s}(k)$ and outputs $\mathbf{ŷ_s}(k)$. $\mathbf{ŷ_s}(k)$ comprises the measured outputs $\mathbf{ŷ_s^m}(k) = \mathbf{y^m}(k) - \mathbf{ŷ_d}(k)$ and unmeasured $\mathbf{ŷ_s^u(k) = 0}$. See [1].

1Desbiens, A., D. Hodouin & É. Plamondon. 2000, "Global predictive control : a unified control structure for decoupling setpoint tracking, feedforward compensation and disturbance rejection dynamics", IEE Proceedings - Control Theory and Applications, vol. 147, no 4, https://doi.org/10.1049/ip-cta:20000443, p. 465–475, ISSN 1350-2379.