Functions: Plant Models

Functions: Plant Models

SimModel

LinModel

NonLinModel

Set Variable Names

Set Operating Points

Linearize

Differential Equation Solvers

The SimModel types represents discrete state-space models that can be used to construct StateEstimators and PredictiveControllers, or as plant simulators by calling evaloutput and updatestate! methods on SimModel instances (to test estimator/controller designs). For time simulations, the states $\mathbf{x}$ are stored inside SimModel instances. Use setstate! method to manually modify them.

Info

The nomenclature in this page introduces the model manipulated input $\mathbf{u}$, measured disturbances $\mathbf{d}$, state $\mathbf{x}$ and output $\mathbf{y}$, four vectors of nu, nd, nx and ny elements, respectively. The $\mathbf{s}$ vector combines the elements of $\mathbf{u}$ and $\mathbf{d}$. See the page Function: Predictive Controllers for a block diagram overview.

SimModel

ModelPredictiveControl.SimModel — Type

abstract type SimModel end

Abstract supertype of LinModel and NonLinModel types.

(model::SimModel)(d=[]) -> y

Functor allowing callable SimModel object as an alias for evaloutput.

Examples

julia> model = NonLinModel((x,u,_,_)->-x + u, (x,_,_)->x .+ 20, 4, 1, 1, 1, solver=nothing);julia> y = model()1-element Vector{Float64}: 20.0

LinModel

ModelPredictiveControl.LinModel — Type

LinModel(sys::StateSpace[, Ts]; i_u=1:size(sys,2), i_d=Int[])

Construct a linear model from state-space model sys with sampling time Ts in seconds.

The system sys can be continuous or discrete-time (Ts can be omitted for the latter). For continuous dynamics, its state-space equations are (discrete case in Extended Help):

\[\begin{aligned} \mathbf{ẋ}(t) &= \mathbf{A x}(t) + \mathbf{B s}(t) \\ \mathbf{y}(t) &= \mathbf{C x}(t) + \mathbf{D s}(t) \end{aligned}\]

with the state $\mathbf{x}$ and output $\mathbf{y}$ vectors. The $\mathbf{s}$ vector comprises the manipulated inputs $\mathbf{u}$ and measured disturbances $\mathbf{d}$, in any order. i_u provides the indices of $\mathbf{s}$ that are manipulated, and i_d, the measured disturbances. The constructor automatically discretizes continuous systems, resamples discrete ones if Ts ≠ sys.Ts, computes a new balancing and minimal state-space realization, and separates the $\mathbf{s}$ terms in two parts (details in Extended Help). The rest of the documentation assumes discrete models since all systems end up in this form.

See also ss

Examples

julia> model1 = LinModel(ss(-0.1, 1.0, 1.0, 0), 2.0) # continuous-time StateSpaceLinModel with a sample time Ts = 2.0 s:└ dimensions:  ├ 1 manipulated inputs u  ├ 1 states x  ├ 1 outputs y  └ 0 measured disturbances djulia> model2 = LinModel(ss(0.4, 0.2, 0.3, 0, 0.1)) # discrete-time StateSpaceLinModel with a sample time Ts = 0.1 s:└ dimensions:  ├ 1 manipulated inputs u  ├ 1 states x  ├ 1 outputs y  └ 0 measured disturbances d

Extended Help

Extended Help

The state-space equations are similar if sys is discrete-time:

\[\begin{aligned} \mathbf{x}(k+1) &= \mathbf{A x}(k) + \mathbf{B s}(k) \\ \mathbf{y}(k) &= \mathbf{C x}(k) + \mathbf{D s}(k) \end{aligned}\]

Continuous dynamics are internally discretized using c2d and :zoh for manipulated inputs, and :tustin, for measured disturbances. Lastly, if sys is discrete and the provided argument Ts ≠ sys.Ts, the system is resampled by using the aforementioned discretization methods.

Note that the constructor transforms the system to its minimal and balancing realization using minreal for controllability/observability. As a consequence, the final state-space representation will be presumably different from the one provided in sys. It is also converted into a more practical form:

\[\begin{aligned} \mathbf{x}(k+1) &= \mathbf{A x}(k) + \mathbf{B_u u}(k) + \mathbf{B_d d}(k) \\ \mathbf{y}(k) &= \mathbf{C x}(k) + \mathbf{D_d d}(k) \end{aligned}\]

Use the syntax LinModel{NT}(A, Bu, C, Bd, Dd, Ts) to force a specific state-space representation.

It is assumed that $\mathbf{D_u=0}$ (or sys is strictly proper) since otherwise the resulting discrete controller would be acausal in its implementation (≠ mathematical causality). Indeed, at each sampling instant $k$, the predictive controller will:

sample an output $\mathbf{y}(k)$ from the plant

computes an action $\mathbf{u}(k)$ and

apply the action on the plant.

There is a causality paradox if $\mathbf{u}(k)$ impacts $\mathbf{y}(k)$ even before computing it.

LinModel(sys::TransferFunction[, Ts]; i_u=1:size(sys,2), i_d=Int[])

Convert to minimal realization state-space when sys is a transfer function.

sys is equal to $\frac{\mathbf{y}(s)}{\mathbf{s}(s)}$ for continuous-time, and $\frac{\mathbf{y}(z)}{\mathbf{s}(z)}$, for discrete-time.

See also tf

Examples

julia> model = LinModel([tf(3, [30, 1]) tf(-2, [5, 1])], 0.5, i_d=[2])LinModel with a sample time Ts = 0.5 s:└ dimensions:  ├ 1 manipulated inputs u  ├ 2 states x  ├ 1 outputs y  └ 1 measured disturbances d

LinModel(sys::DelayLtiSystem, Ts; i_u=1:size(sys,2), i_d=Int[])

Discretize with zero-order hold when sys is a continuous system with delays.

The delays must be multiples of the sample time Ts.

LinModel{NT}(A, Bu, C, Bd, Dd, Ts)

Construct the model from the discrete state-space matrices A, Bu, C, Bd, Dd directly.

See LinModel(::StateSpace) Extended Help for the meaning of the matrices. The arguments Bd and Dd can be the scalar 0 if there is no measured disturbance. This syntax do not modify the state-space representation provided in argument (minreal is not called). Care must be taken to ensure that the model is controllable and observable. The optional parameter NT explicitly set the number type of vectors (default to Float64).

LinModel(model::NonLinModel; x=model.x0+model.xop, u=model.uop, d=model.dop)

Call linearize(model; x, u, d) and return the resulting linear model.

NonLinModel

ModelPredictiveControl.NonLinModel — Type

NonLinModel{NT}(f::Function,  h::Function,  Ts, nu, nx, ny, nd=0; <keyword arguments>)
NonLinModel{NT}(f!::Function, h!::Function, Ts, nu, nx, ny, nd=0; <keyword arguments>)

Construct a nonlinear model from state-space functions f/f! and h/h!.

Both continuous and discrete-time models are supported. The default arguments assume continuous dynamics. Use solver=nothing for the discrete case (see Extended Help). The functions are defined as:

\[\begin{aligned} \mathbf{ẋ}(t) &= \mathbf{f}\Big( \mathbf{x}(t), \mathbf{u}(t), \mathbf{d}(t), \mathbf{p} \Big) \\ \mathbf{y}(t) &= \mathbf{h}\Big( \mathbf{x}(t), \mathbf{d}(t), \mathbf{p} \Big) \end{aligned}\]

where $\mathbf{x}$, $\mathbf{y}$, $\mathbf{u}$, $\mathbf{d}$ and $\mathbf{p}$ are respectively the state, output, manipulated input, measured disturbance and parameter vectors. As a matter of fact, the parameter argument p can be any Julia objects but use a mutable type if you want to change them later e.g.: a vector.

See Extended Help if the dynamics are a function of $t$ or $\mathbf{u}$ appears in $\mathbf{h}$. The functions can be implemented in two possible ways:

Non-mutating functions (out-of-place): define them as f(x, u, d, p) -> ẋ and h(x, d, p) -> y. This syntax is simple and intuitive but it allocates more memory.

Mutating functions (in-place): define them as f!(ẋ, x, u, d, p) -> nothing and h!(y, x, d, p) -> nothing. This syntax reduces the allocations and potentially the computational burden as well.

Tip

Replace the d or p argument with _ in your functions if not needed (see Examples below).

The rest of the documentation assumes discrete dynamics since all models end up in this form. The optional parameter NT explicitly set the number type of vectors (default to Float64).

Warning

The two functions must be in pure Julia to use the model in NonLinMPC, ExtendedKalmanFilter, MovingHorizonEstimator and linearize, except if a finite difference backend is used (e.g. AutoFiniteDiff).

See also LinModel.

Arguments

f::Function or f!: state function of the model.

h::Function or h!: output function of the model.

Ts: sampling time of the model in seconds.

nu: number of manipulated inputs.

nx: number of states.

ny: number of outputs.

nd=0: number of measured disturbances.

p=[]: parameters of the model (any type).

solver=RungeKutta(4): a DiffSolver object for the discretization of continuous dynamics, use nothing for discrete-time models (default to 4th order RungeKutta).

jacobian=AutoForwardDiff(): an AbstractADType backend when linearize is called, see DifferentiationInterface doc.

Examples

julia> f!(ẋ, x, u, _ , p) = (ẋ .= p*x .+ u; nothing);julia> h!(y, x, _ , _ ) = (y .= 0.1x; nothing);julia> model1 = NonLinModel(f!, h!, 5.0, 1, 1, 1, p=-0.2)       # continuous dynamicsNonLinModel with a sample time Ts = 5.0 s:├ solver: RungeKutta(4)├ jacobian: AutoForwardDiff└ dimensions:  ├ 1 manipulated inputs u  ├ 1 states x  ├ 1 outputs y  └ 0 measured disturbances djulia> f(x, u, _ , _ ) = 0.1x + u;julia> h(x, _ , _ ) = 2x;julia> model2 = NonLinModel(f, h, 2.0, 1, 1, 1, solver=nothing) # discrete dynamicsNonLinModel with a sample time Ts = 2.0 s:├ solver: EmptySolver├ jacobian: AutoForwardDiff└ dimensions:  ├ 1 manipulated inputs u  ├ 1 states x  ├ 1 outputs y  └ 0 measured disturbances d

Extended Help

Extended Help

State-space functions are similar for discrete dynamics:

\[\begin{aligned} \mathbf{x}(k+1) &= \mathbf{f}\Big( \mathbf{x}(k), \mathbf{u}(k), \mathbf{d}(k), \mathbf{p} \Big) \\ \mathbf{y}(k) &= \mathbf{h}\Big( \mathbf{x}(k), \mathbf{d}(k), \mathbf{p} \Big) \end{aligned}\]

with two possible implementations as well:

Non-mutating functions: define them as f(x, u, d, p) -> xnext and h(x, d, p) -> y.

Mutating functions: define them as f!(xnext, x, u, d, p) -> nothing and h!(y, x, d, p) -> nothing.

If the dynamics are a function of the time, simply add a measured disturbance defined as $d(t) = t$. This package does not support the $\mathbf{u}$ argument in $\mathbf{h}$ function, see the Extended Help of LinModel for the justification.

Set Variable Names

ModelPredictiveControl.setname! — Function

setname!(model::SimModel; u=nothing, y=nothing, d=nothing, x=nothing) -> model

Set the names of model inputs u, outputs y, disturbances d, and states x.

The keyword arguments u, y, d, and x must be vectors of strings. The strings are used in the plotting functions.

Examples

julia> model = setname!(LinModel(tf(3, [10, 1]), 2.0), u=["\$A\$ (%)"], y=["\$T\$ (∘C)"])LinModel with a sample time Ts = 2.0 s:└ dimensions:  ├ 1 manipulated inputs u  ├ 1 states x  ├ 1 outputs y  └ 0 measured disturbances d

Set Operating Points

ModelPredictiveControl.setop! — Function

setop!(model; uop=nothing, yop=nothing, dop=nothing, xop=nothing, fop=nothing) -> model

Set the operating points of model (both LinModel and NonLinModel).

Introducing deviations vectors around manipulated input uop, model output yop, measured disturbance dop, and model state xop operating points (a.k.a. nominal values):

\[\begin{align*} \mathbf{u_0}(k) &= \mathbf{u}(k) - \mathbf{u_{op}} \\ \mathbf{d_0}(k) &= \mathbf{d}(k) - \mathbf{d_{op}} \\ \mathbf{y_0}(k) &= \mathbf{y}(k) - \mathbf{y_{op}} \\ \mathbf{x_0}(k) &= \mathbf{x}(k) - \mathbf{x_{op}} \\ \end{align*}\]

The state-space description of LinModel around the operating points is:

\[\begin{aligned} \mathbf{x_0}(k+1) &= \mathbf{A x_0}(k) + \mathbf{B_u u_0}(k) + \mathbf{B_d d_0}(k) + \mathbf{f_{op}} - \mathbf{x_{op}} \\ \mathbf{y_0}(k) &= \mathbf{C x_0}(k) + \mathbf{D_d d_0}(k) \end{aligned}\]

and, for NonLinModel:

\[\begin{aligned} \mathbf{x_0}(k+1) &= \mathbf{f}\Big(\mathbf{x_0}(k), \mathbf{u_0}(k), \mathbf{d_0}(k), \mathbf{p}\Big) + \mathbf{f_{op}} - \mathbf{x_{op}} \\ \mathbf{y_0}(k) &= \mathbf{h}\Big(\mathbf{x_0}(k), \mathbf{d_0}(k), \mathbf{p}\Big) \end{aligned}\]

The state xop and the additional fop operating points are frequently zero e.g.: when model comes from system identification. The two vectors are internally used by linearize for non-equilibrium points.

Examples

julia> model = setop!(LinModel(tf(3, [10, 1]), 2.0), uop=[50], yop=[20])LinModel with a sample time Ts = 2.0 s:└ dimensions:  ├ 1 manipulated inputs u  ├ 1 states x  ├ 1 outputs y  └ 0 measured disturbances djulia> y = model()1-element Vector{Float64}: 20.0

Linearize

ModelPredictiveControl.linearize — Function

linearize(model::SimModel; x=model.x0+model.xop, u=model.uop, d=model.dop) -> linmodel

Linearize model at the operating points x, u, d and return the LinModel.

The arguments x, u and d are the linearization points for the state $\mathbf{x}$, manipulated input $\mathbf{u}$ and measured disturbance $\mathbf{d}$, respectively (not necessarily an equilibrium, details in Extended Help). By default, ForwardDiff automatically computes the Jacobians of $\mathbf{f}$ and $\mathbf{h}$ functions. Modify the jacobian keyword argument at the construction of model to swap the backend.

Warning

See Extended Help if you get an error like: MethodError: no method matching (::var"##")(::Vector{ForwardDiff.Dual}).

Examples

julia> model = NonLinModel((x,u,_,_)->x.^3 + u, (x,_,_)->x, 0.1, 1, 1, 1, solver=nothing);julia> linmodel = linearize(model, x=[10.0], u=[0.0]); julia> linmodel.A1×1 Matrix{Float64}: 300.0

Extended Help

Extended Help

With the nonlinear state-space model:

\[\begin{aligned} \mathbf{x}(k+1) &= \mathbf{f}\Big(\mathbf{x}(k), \mathbf{u}(k), \mathbf{d}(k), \mathbf{p}\Big) \\ \mathbf{y}(k) &= \mathbf{h}\Big(\mathbf{x}(k), \mathbf{d}(k), \mathbf{p}\Big) \end{aligned}\]

its linearization at the operating point $\mathbf{x_{op}, u_{op}, d_{op}}$ is:

\[\begin{aligned} \mathbf{x_0}(k+1) &≈ \mathbf{A x_0}(k) + \mathbf{B_u u_0}(k) + \mathbf{B_d d_0}(k) + \mathbf{f(x_{op}, u_{op}, d_{op}, p)} - \mathbf{x_{op}} \\ \mathbf{y_0}(k) &≈ \mathbf{C x_0}(k) + \mathbf{D_d d_0}(k) \end{aligned}\]

based on the deviation vectors $\mathbf{x_0, u_0, d_0, y_0}$ introduced in setop! documentation, and the Jacobian matrices:

\[\begin{aligned} \mathbf{A} &= \left. \frac{∂\mathbf{f(x, u, d, p)}}{∂\mathbf{x}} \right|_{\mathbf{x=x_{op},\, u=u_{op},\, d=d_{op}}} \\ \mathbf{B_u} &= \left. \frac{∂\mathbf{f(x, u, d, p)}}{∂\mathbf{u}} \right|_{\mathbf{x=x_{op},\, u=u_{op},\, d=d_{op}}} \\ \mathbf{B_d} &= \left. \frac{∂\mathbf{f(x, u, d, p)}}{∂\mathbf{d}} \right|_{\mathbf{x=x_{op},\, u=u_{op},\, d=d_{op}}} \\ \mathbf{C} &= \left. \frac{∂\mathbf{h(x, d, p)}}{∂\mathbf{x}} \right|_{\mathbf{x=x_{op},\, d=d_{op}}} \\ \mathbf{D_d} &= \left. \frac{∂\mathbf{h(x, d, p)}}{∂\mathbf{d}} \right|_{\mathbf{x=x_{op},\, d=d_{op}}} \end{aligned}\]

Following setop! notation, we find:

\[\begin{aligned} \mathbf{f_{op}} &= \mathbf{f(x_{op}, u_{op}, d_{op}, p)} \\ \mathbf{y_{op}} &= \mathbf{h(x_{op}, d_{op}, p)} \end{aligned}\]

Notice that $\mathbf{f_{op} - x_{op} = 0}$ if the point is an equilibrium. The equations are similar if the nonlinear model has nonzero operating points.

Automatic differentiation (AD) allows exact Jacobians. The NonLinModel f and h functions must be compatible with this feature though. See JuMP documentation for common mistakes when writing these functions.

ModelPredictiveControl.linearize! — Function

linearize!(linmodel::LinModel, model::SimModel; <keyword arguments>) -> linmodel

Linearize model and store the result in linmodel (in-place).

The keyword arguments are identical to linearize. The code is allocation-free if model simulations does not allocate.

Examples

julia> model = NonLinModel((x,u,_,_)->x.^3 + u, (x,_,_)->x, 0.1, 1, 1, 1, solver=nothing);julia> linmodel = linearize(model, x=[10.0], u=[0.0]); linmodel.A1×1 Matrix{Float64}: 300.0julia> linearize!(linmodel, model, x=[20.0], u=[0.0]); linmodel.A1×1 Matrix{Float64}: 1200.0

Differential Equation Solvers

DiffSolver

ModelPredictiveControl.DiffSolver — Type

Abstract supertype of all differential equation solvers.

RungeKutta

ModelPredictiveControl.RungeKutta — Type

RungeKutta(order=4; supersample=1)

Create an explicit Runge-Kutta solver with optional super-sampling.

The argument order specifies the order of the Runge-Kutta solver, which must be 1 or 4. The keyword argument supersample provides the number of internal steps (default to 1 step). This solver is allocation-free if the f! and h! functions do not allocate.

ForwardEuler

ModelPredictiveControl.ForwardEuler — Function

ForwardEuler(; supersample=1)

Create a Forward Euler solver with optional super-sampling.

This is an alias for RungeKutta(1; supersample), see RungeKutta.

LinModel(sys::StateSpace[, Ts]; i_u=1:size(sys,2), i_d=Int[]) LinModel(sys::TransferFunction[, Ts]; i_u=1:size(sys,2), i_d=Int[]) LinModel(sys::DelayLtiSystem, Ts; i_u=1:size(sys,2), i_d=Int[]) LinModel{NT}(A, Bu, C, Bd, Dd, Ts) LinModel(model::NonLinModel; x=model.x0+model.xop, u=model.uop, d=model.dop)
Construct a linear model from state-space model sys with sampling time Ts in seconds.

LinModel(sys::StateSpace[, Ts]; i_u=1:size(sys,2), i_d=Int[]) LinModel(sys::TransferFunction[, Ts]; i_u=1:size(sys,2), i_d=Int[]) LinModel{NT}(A, Bu, C, Bd, Dd, Ts) LinModel(model::NonLinModel; x=model.x0+model.xop, u=model.uop, d=model.dop)
Construct a linear model from state-space model sys with sampling time Ts in seconds.

LinModel(sys::StateSpace[, Ts]; i_u=1:size(sys,2), i_d=Int[]) LinModel(sys::TransferFunction[, Ts]; i_u=1:size(sys,2), i_d=Int[]) LinModel(sys::DelayLtiSystem, Ts; i_u=1:size(sys,2), i_d=Int[])
Construct a linear model from state-space model sys with sampling time Ts in seconds.

NonLinModel{NT}(f::Function, h::Function, Ts, nu, nx, ny, nd=0; <keyword arguments>) NonLinModel{NT}(f!::Function, h!::Function, Ts, nu, nx, ny, nd=0; <keyword arguments>)
Construct a nonlinear model from state-space functions f/f! and h/h!.

abstract type SimModel end
Abstract supertype of LinModel and NonLinModel types.

linearize(model::SimModel; x=model.x0+model.xop, u=model.uop, d=model.dop) -> linmodel
Linearize model at the operating points x, u, d and return the LinModel.

f!(x0next, _ , model::LinModel, x0, u0, d0, _ ) -> nothing f!(x0next, k, model::NonLinModel, x0, u0, d0, p)
Evaluate x0next = A*x0 + Bu*u0 + Bd*d0 in-place when model is a LinModel.

h!(y0, model::LinModel, x0, d0, _ ) -> nothing h!(y0, model::NonLinModel, x0, d0, p)
Evaluate y0 = C*x0 + Dd*d0 in-place when model is a LinModel.