State Estimators

State Estimators

StateEstimator

SteadyKalmanFilter

KalmanFilter

UnscentedKalmanFilter

InternalModel

This module includes many state estimators (or state observer), both for deterministic and stochastic systems. The implementations focus on control applications, that is, relying on the estimates to compute a full state feedback (predictive controllers, in this package). They all incorporates some kind of integral action by default, since it is generally desired to eliminate the steady-state error with closed-loop control (offset-free tracking).

Info

If you plan to use the estimators for other contexts than this specific package (e.g. : filter, parameter estimation, etc.), careful must be taken at construction since the integral action is not necessarily desired. The option nint_ym=0 disable it.

The estimators are all implemented in the predictor form (a.k.a. observer form), that is, they all estimates at each discrete time $k$ the states of the next period $\mathbf{x̂}_k(k+1)$. In contrast, the filter form that estimates $\mathbf{x̂}_k(k)$ is sometimes slightly more accurate.

The predictor form comes in handy for control applications since the estimations come after the controller computations, without introducing any additional delays. Moreover, the moveinput! method of the predictive controllers does not automatically update the estimates with updatestate!. This allows applying the calculated inputs on the real plant before starting the potentially expensive estimator computations (see Manual for examples).

Info

All the estimators support measured $\mathbf{y^m}$ and unmeasured $\mathbf{y^u}$ model outputs, where $\mathbf{y}$ refers to all of them.

StateEstimator

ModelPredictiveControl.StateEstimator — Type

Abstract supertype of all state estimators.

(estim::StateEstimator)(d=Float64[])

Functor allowing callable StateEstimator object as an alias for evaloutput.

Examples

julia> kf = KalmanFilter(setop!(LinModel(tf(3, [10, 1]), 2), yop=[20]));

julia> ŷ = kf() 
1-element Vector{Float64}:
 20.0

SteadyKalmanFilter

ModelPredictiveControl.SteadyKalmanFilter — Type

SteadyKalmanFilter(model::LinModel; <keyword arguments>)

Construct a steady-state Kalman Filter with the LinModel model.

The steady-state (or asymptotic) Kalman filter is based on the process model :

\[\begin{aligned} \mathbf{x}(k+1) &= \mathbf{Â x}(k) + \mathbf{B̂_u u}(k) + \mathbf{B̂_d d}(k) + \mathbf{w}(k) \\ \mathbf{y^m}(k) &= \mathbf{Ĉ^m x}(k) + \mathbf{D̂_d^m d}(k) + \mathbf{v}(k) \\ \mathbf{y^u}(k) &= \mathbf{Ĉ^u x}(k) + \mathbf{D̂_d^u d}(k) \end{aligned}\]

with sensor $\mathbf{v}(k)$ and process $\mathbf{w}(k)$ noises as uncorrelated zero mean white noise vectors, with a respective covariance of $\mathbf{R̂}$ and $\mathbf{Q̂}$. The arguments are in standard deviations σ, i.e. same units than outputs and states. The matrices $\mathbf{Â, B̂_u, B̂_d, Ĉ, D̂_d}$ are model matrices augmented with the stochastic model, which is specified by the numbers of output integrator nint_ym (see Extended Help). Likewise, the covariance matrices are augmented with $\mathbf{Q̂ = \text{diag}(Q, Q_{int})}$ and $\mathbf{R̂ = R}$. The matrices $\mathbf{Ĉ^m, D̂_d^m}$ are the rows of $\mathbf{Ĉ, D̂_d}$ that correspond to measured outputs $\mathbf{y^m}$ (and unmeasured ones, for $\mathbf{Ĉ^u, D̂_d^u}$).

Arguments

model::LinModel : (deterministic) model for the estimations.

i_ym=1:model.ny : model output indices that are measured $\mathbf{y^m}$, the rest are unmeasured $\mathbf{y^u}$.

σQ=fill(0.1,model.nx) : main diagonal of the process noise covariance $\mathbf{Q}$ of model, specified as a standard deviation vector.

σR=fill(0.1,length(i_ym)) : main diagonal of the sensor noise covariance $\mathbf{R}$ of model measured outputs, specified as a standard deviation vector.

nint_ym=fill(1,length(i_ym)) : integrator quantity per measured outputs (vector) for the stochastic model, use nint_ym=0 for no integrator at all.

σQ_int=fill(0.1,sum(nint_ym)) : same than σQ but for the stochastic model covariance $\mathbf{Q_{int}}$ (composed of output integrators).

Examples

julia> model = LinModel([tf(3, [30, 1]); tf(-2, [5, 1])], 0.5);

julia> estim = SteadyKalmanFilter(model, i_ym=[2], σR=[1], σQ_int=[0.01])
SteadyKalmanFilter estimator with a sample time Ts = 0.5 s and:
 1 manipulated inputs u
 3 states x̂
 1 measured outputs ym
 1 unmeasured outputs yu
 0 measured disturbances d

Extended Help

The model augmentation with nint_ym vector produces the integral action when the estimator is used in a controller as state feedback. The default is 1 integrator per measured outputs, resulting in an offset-free tracking for "step-like" unmeasured output disturbances. Use 2 integrators for "ramp-like" disturbances. See init_estimstoch.

Tip

Increasing σQ_int values increases the integral action "gain".

The constructor pre-compute the steady-state Kalman gain K with the kalman function. It can sometimes fail, for example when model is integrating. In such a case, you can use 0 integrator on model integrating outputs, or the alternative time-varying KalmanFilter.

KalmanFilter

ModelPredictiveControl.KalmanFilter — Type

KalmanFilter(model::LinModel; <keyword arguments>)

Construct a time-varying Kalman Filter with the LinModel model.

The process model is identical to SteadyKalmanFilter. The matrix $\mathbf{P̂}_k(k+1)$ is the estimation error covariance of model states augmented with the stochastic ones (specified by nint_ym). Two keyword arguments modify its initial value with $\mathbf{P̂}_{-1}(0) = \mathrm{diag}\{ \mathbf{P}(0), \mathbf{P_{int}}(0) \}$.

Arguments

model::LinModel : (deterministic) model for the estimations.

σP0=fill(10,model.nx) : main diagonal of the initial estimate covariance $\mathbf{P}(0)$, specified as a standard deviation vector.

σP0_int=fill(10,sum(nint_ym)) : same than σP0 but for the stochastic model covariance $\mathbf{P_{int}}(0)$ (composed of output integrators).

<keyword arguments> of SteadyKalmanFilter constructor.

Examples

julia> model = LinModel([tf(3, [30, 1]); tf(-2, [5, 1])], 0.5);

julia> estim = KalmanFilter(model, i_ym=[2], σR=[1], σP0=[100, 100], σQ_int=[0.01])
KalmanFilter estimator with a sample time Ts = 0.5 s and:
 1 manipulated inputs u
 3 states x̂
 1 measured outputs ym
 1 unmeasured outputs yu
 0 measured disturbances d

UnscentedKalmanFilter

ModelPredictiveControl.UnscentedKalmanFilter — Type

UnscentedKalmanFilter(model::SimModel; <keyword arguments>)

Construct an unscented Kalman Filter with the SimModel model.

Both LinModel and NonLinModel are supported. The unscented Kalman filter is based on the process model :

\[\begin{aligned} \mathbf{x}(k+1) &= \mathbf{f̂}\Big(\mathbf{x̂}(k), \mathbf{u}(k), \mathbf{d}(k)\Big) + \mathbf{w}(k) \\ \mathbf{y^m}(k) &= \mathbf{ĥ^m}\Big(\mathbf{x̂}(k), \mathbf{d}(k)\Big) + \mathbf{v}(k) \\ \mathbf{y^u}(k) &= \mathbf{ĥ^u}\Big(\mathbf{x̂}(k), \mathbf{d}(k)\Big) \\ \end{aligned}\]

See SteadyKalmanFilter for details on $\mathbf{v}(k), \mathbf{w}(k)$ noises and $\mathbf{R̂}, \mathbf{Q̂}$ covariances. The functions $\mathbf{f̂, ĥ}$ are model state-space functions augmented with the stochastic model, which is specified by the numbers of output integrator nint_ym (see SteadyKalmanFilter for details). The $\mathbf{ĥ^m}$ function represents the measured outputs of $\mathbf{ĥ}$ function (and unmeasured ones, for $\mathbf{ĥ^u}$).

Arguments

model::SimModel : (deterministic) model for the estimations.

α=1e-3 : alpha parameter, spread of the state distribution $(0 ≤ α ≤ 1)$.

β=2 : beta parameter, skewness and kurtosis of the states distribution $(β ≥ 0)$.

κ=0 : kappa parameter, another spread parameter $(0 ≤ κ ≤ 3)$.

<keyword arguments> of SteadyKalmanFilter constructor.

<keyword arguments> of KalmanFilter constructor.

Examples

julia> model = NonLinModel((x,u,_)->0.1x+u, (x,_)->2x, 10, 1, 1, 1);

julia> estim = UnscentedKalmanFilter(model, σR=[1], nint_ym=[2], σP0_int=[1, 1])
UnscentedKalmanFilter{NonLinModel} estimator with a sample time Ts = 10.0 s and:
 1 manipulated inputs u
 3 states x̂
 1 measured outputs ym
 0 unmeasured outputs yu
 0 measured disturbances d

InternalModel

ModelPredictiveControl.InternalModel — Type

InternalModel(model::SimModel; i_ym=1:model.ny, stoch_ym=ss(1,1,1,1,model.Ts).*I)

Construct an internal model estimator based on model (LinModel or NonLinModel).

i_ym provides the model output indices that are measured $\mathbf{y^m}$, the rest are unmeasured $\mathbf{y^u}$. model evaluates the deterministic predictions $\mathbf{ŷ_d}$, and stoch_ym, the stochastic predictions of the measured outputs $\mathbf{ŷ_s^m}$ (the unmeasured ones being $\mathbf{ŷ_s^u=0}$). The predicted outputs sum both values : $\mathbf{ŷ = ŷ_d + ŷ_s}$.

Warning

InternalModel estimator does not work if model is integrating or unstable. The constructor verifies these aspects for LinModel but not for NonLinModel. Uses any other state estimator in such cases.

Examples

julia> estim = InternalModel(LinModel([tf(3, [30, 1]); tf(-2, [5, 1])], 0.5), i_ym=[2])
InternalModel{LinModel} estimator with a sample time Ts = 0.5 s and:
 1 manipulated inputs u
 2 states x̂
 1 measured outputs ym
 1 unmeasured outputs yu
 0 measured disturbances d

Extended Help

stoch_ym is a TransferFunction or StateSpace object that models disturbances on $\mathbf{y^m}$. Its input is a hypothetical zero mean white noise vector. stoch_ym supposes 1 integrator per measured outputs by default, assuming that the current stochastic estimate $\mathbf{ŷ_s^m}(k) = \mathbf{y^m}(k) - \mathbf{ŷ_d^m}(k)$ is constant in the future. This is the dynamic matrix control (DMC) strategy, which is simple but sometimes too aggressive. Additional poles and zeros in stoch_ym can mitigate this.