ModelPredictiveControl.jl Documentation

ModelPredictiveControl.LinModel — Type

LinModel(sys::StateSpace, Ts=NaN; i_u=1:size(sys,2), i_d=Int[])

Construct a LinModel from state-space model sys with sampling time Ts in second.

Ts can be omitted when sys is discrete-time. Its state-space matrices are:

\[\begin{align*} \mathbf{x}(k+1) &= \mathbf{A} \mathbf{x}(k) + \mathbf{B} \mathbf{z}(k) \\ \mathbf{y}(k) &= \mathbf{C} \mathbf{x}(k) + \mathbf{D} \mathbf{z}(k) \end{align*}\]

with the state $\mathbf{x}$ and output $\mathbf{y}$ vectors. The $\mathbf{z}$ vector comprises the manipulated inputs $\mathbf{u}$ and measured disturbances $\mathbf{d}$, in any order. i_u provides the indices of $\mathbf{z}$ that are manipulated, and i_d, the measured disturbances. The state-space matrices are similar if sys is continuous-time (replace $\mathbf{x}(k+1)$ with $\dot{\mathbf{x}}(t)$). In such a case, it's discretized with c2d and :zoh for manipulated inputs, and :tustin, for measured disturbances.

The constructor transforms the system to a more practical form (Dᵤ = 0 because of the zero-order hold):

\[\begin{align*} \mathbf{x}(k+1) &= \mathbf{A} \mathbf{x}(k) + \mathbf{B_u} \mathbf{u}(k) + \mathbf{B_d} \mathbf{d}(k) \\ \mathbf{y}(k) &= \mathbf{C} \mathbf{x}(k) + \mathbf{D_d} \mathbf{d}(k) \end{align*}\]

See also ss, tf.

Examples

julia> model = LinModel(ss(0.4, 0.2, 0.3, 0, 0.1))
Discrete-time linear model with a sample time Ts = 0.1 s and:
- 1 manipulated inputs u
- 1 states x
- 1 outputs y
- 0 measured disturbances d

LinModel(sys::TransferFunction, Ts=NaN; i_u=1:size(sys,2), i_d=Int[])

Convert to minimal realization state-space when sys is a transfer function.

sys is equal to $\frac{\mathbf{y}(s)}{\mathbf{z}(s)}$ for continuous-time, and $\frac{\mathbf{y}(z)}{\mathbf{z}(z)}$, for discrete-time.

Examples

julia> model = LinModel([tf(3, [30, 1]) tf(-2, [5, 1])], 2, i_d=[2])
Discrete-time linear model with a sample time Ts = 2.0 s and:
- 1 manipulated inputs u
- 2 states x
- 1 outputs y
- 1 measured disturbances d

ModelPredictiveControl.NonLinModel — Type

NonLinModel(f, h, Ts::Real, nu::Int, nx::Int, ny::Int, nd::Int=0)

Construct a NonLinModel from discrete-time state-space functions f and h.

The state update $\mathbf{f}$ and output $\mathbf{h}$ functions are defined as :

\[ \begin{align*} \mathbf{x}(k+1) &= \mathbf{f}\Big( \mathbf{x}(k), \mathbf{u}(k), \mathbf{d}(k) \Big) \\ \mathbf{y}(k) &= \mathbf{h}\Big( \mathbf{x}(k), \mathbf{d}(k) \Big) \end{align*}\]

Ts is the sampling time in second. nu, nx, ny and nd are the respective number of manipulated inputs, states, outputs and measured disturbances. Replace the d argument with _ if nd=0 (see Examples below). Nonlinear continuous-time state-space functions are not supported for the time being. In such a case, manually call a differential equation solver in the f function (e.g.: Euler method).

See also LinModel.

Examples

julia> model = NonLinModel((x,u,_)->-x+u, (x,_)->2x, 10, 1 , 1 , 1)
Discrete-time nonlinear model with a sample time Ts = 10.0 s and:
- 1 manipulated inputs u
- 1 states x
- 1 outputs y
- 0 measured disturbances d

ModelPredictiveControl.setop! — Function

setop!(model::SimModel; uop=Float64[], yop=Float64[], dop=Float64[])

Set model inputs uop, outputs yop and measured disturbances dop operating points.

The state-space model including operating points (a.k.a. nominal values) is:

\[\begin{align*} \mathbf{x}(k+1) &= \mathbf{A} \mathbf{x}(k) + \mathbf{B_u} \mathbf{u_0}(k) + \mathbf{B_d} \mathbf{d_0}(k) \\ \mathbf{y_0}(k) &= \mathbf{C} \mathbf{x}(k) + \mathbf{D_d} \mathbf{d_0}(k) \end{align*}\]

where

\[\begin{align*} \mathbf{u_0}(k) &= \mathbf{u}(k) - \mathbf{u_{op}}(k) \\ \mathbf{y_0}(k) &= \mathbf{y}(k) - \mathbf{y_{op}}(k) \\ \mathbf{d_0}(k) &= \mathbf{d}(k) - \mathbf{d_{op}}(k) \end{align*}\]

The structure is similar if model is a NonLinModel:

\[\begin{align*} \mathbf{x}(k+1) &= \mathbf{f}\Big(\mathbf{x}(k), \mathbf{u_0}(k), \mathbf{d_0}(k)\Big)\\ \mathbf{y_0}(k) &= \mathbf{h}\Big(\mathbf{x}(k), \mathbf{d_0}(k)\Big) \end{align*}\]

Examples

julia> model = LinModel(tf(3, [10, 1]), 2);

julia> setop!(model, uop=[50], yop=[20])

ModelPredictiveControl.updatestate — Function

updatestate(sys::SimModel, x, u, d=Float64[])

Update states x of sys with current inputs u and measured disturbances d.

ModelPredictiveControl.evaloutput — Function

evaloutput(sys::SimModel, x, d=Float64[])

Evaluate output y of sys with current state x and measured disturbances d.

Missing docstring.

Missing docstring for init_internalmodel. Check Documenter's build log for details.