ModelPredictiveControl.jl Documentation

ModelPredictiveControl.jl Documentation

Tutorial

Specifying models

SimModel functions

State estimator design

StateEstimator functions

Advanced Topics

Internals

API

Tutorial

Specifying models

adsasd

SimModel functions

ModelPredictiveControl.LinModel — Type

LinModel(sys::StateSpace[, Ts]; i_u=1:size(sys,2), i_d=Int[])

Construct a LinModel from state-space model sys with sampling time Ts in second.

Ts can be omitted when sys is discrete-time. Its state-space matrices are:

\[\begin{aligned} \mathbf{x}(k+1) &= \mathbf{A x}(k) + \mathbf{B z}(k) \\ \mathbf{y}(k) &= \mathbf{C x}(k) + \mathbf{D z}(k) \end{aligned}\]

with the state $\mathbf{x}$ and output $\mathbf{y}$ vectors. The $\mathbf{z}$ vector comprises the manipulated inputs $\mathbf{u}$ and measured disturbances $\mathbf{d}$, in any order. i_u provides the indices of $\mathbf{z}$ that are manipulated, and i_d, the measured disturbances. The state-space matrices are similar if sys is continuous-time (replace x(k+1) with ẋ(t)). In such a case, it's discretized with c2d and :zoh for manipulated inputs, and :tustin, for measured disturbances.

The constructor transforms the system to a more practical form (Dᵤ = 0 because of the zero-order hold):

\[\begin{aligned} \mathbf{x}(k+1) &= \mathbf{A x}(k) + \mathbf{B_u u}(k) + \mathbf{B_d d}(k) \\ \mathbf{y}(k) &= \mathbf{C x}(k) + \mathbf{D_d d}(k) \end{aligned}\]

See also ss, tf.

Examples

julia> model = LinModel(ss(0.4, 0.2, 0.3, 0, 0.1))
Discrete-time linear model with a sample time Ts = 0.1 s and:
 1 manipulated inputs u
 1 states x
 1 outputs y
 0 measured disturbances d

LinModel(sys::TransferFunction[, Ts]; i_u=1:size(sys,2), i_d=Int[])

Convert to minimal realization state-space when sys is a transfer function.

sys is equal to $\frac{\mathbf{y}(s)}{\mathbf{z}(s)}$ for continuous-time, and $\frac{\mathbf{y}(z)}{\mathbf{z}(z)}$, for discrete-time.

Examples

julia> model = LinModel([tf(3, [30, 1]) tf(-2, [5, 1])], 0.5, i_d=[2])
Discrete-time linear model with a sample time Ts = 0.5 s and:
 1 manipulated inputs u
 2 states x
 1 outputs y
 1 measured disturbances d

ModelPredictiveControl.NonLinModel — Type

NonLinModel(f, h, Ts::Real, nu::Int, nx::Int, ny::Int, nd::Int=0)

Construct a NonLinModel from discrete-time state-space functions f and h.

The state update $\mathbf{f}$ and output $\mathbf{h}$ functions are defined as :

\[ \begin{aligned} \mathbf{x}(k+1) &= \mathbf{f}\Big( \mathbf{x}(k), \mathbf{u}(k), \mathbf{d}(k) \Big) \\ \mathbf{y}(k) &= \mathbf{h}\Big( \mathbf{x}(k), \mathbf{d}(k) \Big) \end{aligned}\]

Ts is the sampling time in second. nu, nx, ny and nd are the respective number of manipulated inputs, states, outputs and measured disturbances.

Tip

Replace the d argument with _ if nd = 0 (see Examples below).

Nonlinear continuous-time state-space functions are not supported for now. In such a case, manually call a differential equation solver in the f function (e.g.: Euler method).

See also LinModel.

Examples

julia> model = NonLinModel((x,u,_)->-x+u, (x,_)->2x, 10, 1 , 1 , 1)
Discrete-time nonlinear model with a sample time Ts = 10.0 s and:
 1 manipulated inputs u
 1 states x
 1 outputs y
 0 measured disturbances d

ModelPredictiveControl.setop! — Function

setop!(model::SimModel; uop=Float64[], yop=Float64[], dop=Float64[])

Set model inputs uop, outputs yop and measured disturbances dop operating points.

The state-space model including operating points (a.k.a. nominal values) is:

\[\begin{aligned} \mathbf{x}(k+1) &= \mathbf{A x}(k) + \mathbf{B_u u_0}(k) + \mathbf{B_d d_0}(k) \\ \mathbf{y_0}(k) &= \mathbf{C x}(k) + \mathbf{D_d d_0}(k) \end{aligned}\]

where

\[\begin{aligned} \mathbf{u_0}(k) &= \mathbf{u}(k) - \mathbf{u_{op}}(k) \\ \mathbf{y_0}(k) &= \mathbf{y}(k) - \mathbf{y_{op}}(k) \\ \mathbf{d_0}(k) &= \mathbf{d}(k) - \mathbf{d_{op}}(k) \end{aligned}\]

The structure is similar if model is a NonLinModel:

\[\begin{aligned} \mathbf{x}(k+1) &= \mathbf{f}\Big(\mathbf{x}(k), \mathbf{u_0}(k), \mathbf{d_0}(k)\Big)\\ \mathbf{y_0}(k) &= \mathbf{h}\Big(\mathbf{x}(k), \mathbf{d_0}(k)\Big) \end{aligned}\]

Examples

julia> model = setop!(LinModel(tf(3, [10, 1]), 2), uop=[50], yop=[20])
Discrete-time linear model with a sample time Ts = 2.0 s and:
 1 manipulated inputs u
 1 states x
 1 outputs y
 0 measured disturbances d

ModelPredictiveControl.updatestate! — Function

Update states x in model with current inputs u and measured disturbances d.

Update estim.state values with current inputs u, measured outputs ym and dist. d.

ModelPredictiveControl.evaloutput — Function

Evaluate output y of model with current states x and measured disturbances d.

evaloutput(estim::InternalModel, d=Float64[], ym=nothing)

Evaluate InternalModel outputs ̂ŷ from estim.state values and measured outputs ym.

InternalModel estimator needs current measured outputs $\mathbf{y^m}(k)$ to estimate its current outputs $\mathbf{ŷ^m}(k)$, since the strategy imposes that $\mathbf{ŷ^m}(k) = \mathbf{ŷ^m}(k)$ is always true.

State estimator design

Info

All the state estimators support measured $\mathbf{y^m}$ and unmeasured $\mathbf{y^u}$ model outputs, where $\mathbf{y}$ refers to all of them.

StateEstimator functions

ModelPredictiveControl.InternalModel — Type

InternalModel(model::SimModel; i_ym=1:model.ny, stoch_ym=ss(1,1,1,1,model.Ts).*I)

Construct an InternalModel estimator based on model.

i_ym provides the model output indices that are measured $\mathbf{y^m}$, the rest are unmeasured $\mathbf{y^u}$. model evaluates the deterministic predictions $\mathbf{ŷ_d}$, and stoch_ym, the stochastic predictions of the measured outputs $\mathbf{ŷ_s^m}$ (the unmeasured ones being $\mathbf{ŷ_s^u} = \mathbf{0}$).

stoch_ym is a TransferFunction or StateSpace model that hypothetically filters a zero mean white noise vector. Its default value supposes 1 integrator per measured outputs, assuming that the current stochastic estimate $\mathbf{ŷ_s^m}(k) = \mathbf{y^m}(k) - \mathbf{ŷ_d^m}(k)$ will be constant in the future. This is the dynamic matrix control (DMC) strategy, which is simple but sometimes too aggressive. Additional poles and zeros in stoch_ym can mitigate this.

Warning

InternalModel estimator does not work if model is integrating or unstable. The constructor verifies these aspects for LinModel but not for NonLinModel. Uses any other state estimator in such cases.

See also init_internalmodel

Examples

julia> estim = InternalModel(LinModel([tf(3, [30, 1]); tf(-2, [5, 1])], 0.5), i_ym=[2])
InternalModel state estimator with a sample time Ts = 0.5 s and:
 1 manipulated inputs u
 2 states x̂
 1 measured outputs ym
 1 unmeasured outputs yu
 0 measured disturbances d

Advanced Topics

Internals

ModelPredictiveControl.init_internalmodel — Function

init_internalmodel(As, Bs, Cs, Ds)

Calc stochastic model update matrices Âs and B̂s for InternalModel estimator.

Âs and B̂s are the stochastic model update matrices :

\[ \mathbf{x̂_s}(k+1) = \mathbf{Â_s x̂_s}(k) + \mathbf{B̂_s ŷ_s}(k)\]

with current stochastic model states $\mathbf{x̂_s}(k)$ and outputs $\mathbf{ŷ_s}(k)$, which is in turn composed of the measured $\mathbf{ŷ_s^m}(k) = \mathbf{y^m}(k) - \mathbf{ŷ_d^m}(k)$ and unmeasured $\mathbf{ŷ_s^u = 0}$ outputs. See [1].

API

ModelPredictiveControl.InternalModel

ModelPredictiveControl.LinModel

ModelPredictiveControl.NonLinModel

ModelPredictiveControl.evaloutput

ModelPredictiveControl.init_internalmodel

ModelPredictiveControl.setop!

ModelPredictiveControl.updatestate!

1Desbiens, A., D. Hodouin & É. Plamondon. 2000, "Global predictive control : a unified control structure for decoupling setpoint tracking, feedforward compensation and disturbance rejection dynamics", IEE Proceedings - Control Theory and Applications, vol. 147, no 4, https://doi.org/10.1049/ip-cta:20000443, p. 465–475, ISSN 1350-2379.