PredictiveController Internals

The prediction methodology of this module is mainly based on Maciejowski textbook [1].

ModelPredictiveControl.init_deterpred — Function

init_deterpred(model::LinModel, Hp, Hc)

Construct deterministic prediction matrices for LinModel model.

The linear model predictions are evaluated by :

\[\begin{aligned} \mathbf{Ŷ} &= \mathbf{E ΔU} + \mathbf{G d}(k) + \mathbf{J D̂} + \mathbf{K_d x̂_d}(k) + \mathbf{Q u}(k-1) + \mathbf{Ŷ_s} \\ &= \mathbf{E ΔU} + \mathbf{F} \end{aligned}\]

where predicted outputs $\mathbf{Ŷ}$, stochastic outputs $\mathbf{Ŷ_s}$, and measured disturbances $\mathbf{D̂}$ are from $k + 1$ to $k + H_p$. Input increments $\mathbf{ΔU}$ are from $k$ to $k + H_c - 1$. Deterministic state estimates $\mathbf{x̂_d}(k)$ are extracted from current estimates $\mathbf{x̂}_{k-1}(k)$ with split_state. Operating points on $\mathbf{u}$, $\mathbf{d}$ and $\mathbf{y}$ are omitted in above equations.

Extended Help

Using the $\mathbf{A, B_u, C, B_d, D_d}$ matrices in model and the equation $\mathbf{W}_j = \mathbf{C} ( ∑_{i=0}^j \mathbf{A}^i ) \mathbf{B_u}$, the prediction matrices are computed by :

\[\begin{aligned} \mathbf{E} &= \begin{bmatrix} \mathbf{W}_{0} & \mathbf{0} & \cdots & \mathbf{0} \\ \mathbf{W}_{1} & \mathbf{W}_{0} & \cdots & \mathbf{0} \\ \vdots & \vdots & \ddots & \vdots \\ \mathbf{W}_{H_p-1} & \mathbf{W}_{H_p-2} & \cdots & \mathbf{W}_{H_p-H_c+1} \end{bmatrix} \\ \mathbf{G} &= \begin{bmatrix} \mathbf{C}\mathbf{A}^{0} \mathbf{B_d} \\ \mathbf{C}\mathbf{A}^{1} \mathbf{B_d} \\ \vdots \\ \mathbf{C}\mathbf{A}^{H_p-1} \mathbf{B_d} \end{bmatrix} \\ \mathbf{J} &= \begin{bmatrix} \mathbf{D_d} & \mathbf{0} & \cdots & \mathbf{0} \\ \mathbf{C}\mathbf{A}^{0} \mathbf{B_d} & \mathbf{D_d} & \cdots & \mathbf{0} \\ \vdots & \vdots & \ddots & \vdots \\ \mathbf{C}\mathbf{A}^{H_p-2} \mathbf{B_d} & \mathbf{C}\mathbf{A}^{H_p-3} \mathbf{B_d} & \cdots & \mathbf{D_d} \end{bmatrix} \\ \mathbf{K_d} &= \begin{bmatrix} \mathbf{C}\mathbf{A}^{0} \\ \mathbf{C}\mathbf{A}^{1} \\ \vdots \\ \mathbf{C}\mathbf{A}^{H_p-1} \end{bmatrix} \\ \mathbf{Q} &= \begin{bmatrix} \mathbf{W}_0 \\ \mathbf{W}_1 \\ \vdots \\ \mathbf{W}_{H_p-1} \end{bmatrix} \end{aligned}\]

Note

Stochastic predictions $\mathbf{Ŷ_s}$ are calculated separately (see init_stochpred) and added to the $\mathbf{F}$ matrix to support internal model structure and reduce NonLinMPC computational costs. That is also why the prediction matrices are built on $\mathbf{A, B_u, C, B_d, D_d}$ instead of the augmented model $\mathbf{Â, B̂_u, Ĉ, B̂_d, D̂_d}$.

Return empty matrices if model is not a LinModel

ModelPredictiveControl.init_ΔUtoU — Function

init_ΔUtoU(nu, Hp, Hc, C, c_Umin, c_Umax)

Init manipulated input increments to inputs conversion matrices.

The conversion from the input increments $\mathbf{ΔU}$ to manipulated inputs over $H_p$ and $H_c$ are calculated by:

\[\begin{aligned} \mathbf{U} = \mathbf{U}_{H_p} &= \mathbf{S}_{H_p} \mathbf{ΔU} + \mathbf{T}_{H_p} \mathbf{u}(k-1) \\ \mathbf{U}_{H_c} &= \mathbf{S}_{H_c} \mathbf{ΔU} + \mathbf{T}_{H_c} \mathbf{u}(k-1) \end{aligned}\]

ModelPredictiveControl.relaxU — Function

relaxU(C, c_Umin, c_Umax, S_Hp, S_Hc)

Augment manipulated inputs constraints with slack variable ϵ for softening.

Denoting the input increments augmented with the slack variable $\mathbf{ΔŨ} = [\begin{smallmatrix} \mathbf{ΔU} \\ ϵ \end{smallmatrix}]$, it returns the augmented conversion matrices $\mathbf{S̃}_{H_p}$ and $\mathbf{S̃}_{H_c}$, similar to the ones described at init_ΔUtoU. It also returns the $\mathbf{A}$ matrices for the inequality constraints:

\[\begin{bmatrix} \mathbf{A_{U_{min}}} \\ \mathbf{A_{U_{max}}} \end{bmatrix} \mathbf{ΔŨ} ≤ \begin{bmatrix} - \mathbf{U_{min}} + \mathbf{T}_{H_c} \mathbf{u}(k-1) \\ + \mathbf{U_{max}} - \mathbf{T}_{H_c} \mathbf{u}(k-1) \end{bmatrix}\]

ModelPredictiveControl.relaxΔU — Function

relaxΔU(C, c_ΔUmin, c_ΔUmax, ΔUmin, ΔUmax, N_Hc)

Augment input increments constraints with slack variable ϵ for softening.

Denoting the input increments augmented with the slack variable $\mathbf{ΔŨ} = [\begin{smallmatrix} \mathbf{ΔU} \\ ϵ \end{smallmatrix}]$, it returns the augmented input increment weights $\mathbf{Ñ}_{H_c}$ (that incorporate $C$). It also returns the augmented constraints $\mathbf{ΔŨ_{min}}$ and $\mathbf{ΔŨ_{max}}$ and the $\mathbf{A}$ matrices for the inequality constraints:

\[\begin{bmatrix} \mathbf{A_{ΔŨ_{min}}} \\ \mathbf{A_{ΔŨ_{max}}} \end{bmatrix} \mathbf{ΔŨ} ≤ \begin{bmatrix} - \mathbf{ΔŨ_{min}} \\ + \mathbf{ΔŨ_{max}} \end{bmatrix}\]

ModelPredictiveControl.relaxŶ — Function

relaxŶ(::LinModel, C, c_Ŷmin, c_Ŷmax, E)

Augment linear output prediction constraints with slack variable ϵ for softening.

Denoting the input increments augmented with the slack variable $\mathbf{ΔŨ} = [\begin{smallmatrix} \mathbf{ΔU} \\ ϵ \end{smallmatrix}]$, it returns the $\mathbf{Ẽ}$ matrix that appears in the linear model prediction equation $\mathbf{Ŷ = Ẽ ΔŨ + F}$, and the $\mathbf{A}$ matrices for the inequality constraints:

\[\begin{bmatrix} \mathbf{A_{Ŷ_{min}}} \\ \mathbf{A_{Ŷ_{max}}} \end{bmatrix} \mathbf{ΔŨ} ≤ \begin{bmatrix} - \mathbf{Ŷ_{min}} + \mathbf{F} \\ + \mathbf{Ŷ_{max}} - \mathbf{F} \end{bmatrix}\]

Return empty matrices for models different than LinModel

ModelPredictiveControl.init_quadprog — Function

init_quadprog(Ẽ, S_Hp, M_Hp, N_Hc, L_Hp)

Init the quadratic programming optimization matrix P̃.

The P̃ matrix appears in the quadratic general form :

\[ J = \min_{\mathbf{ΔŨ}} \frac{1}{2}\mathbf{(ΔŨ)'P̃(ΔŨ)} + \mathbf{q̃'(ΔŨ)} + p \]

$\mathbf{P̃}$ is constant if the model and weights are linear and time invariant (LTI). The vector $\mathbf{q̃}$ and scalar $p$ need recalculation each control period $k$ (see init_prediction method). $p$ does not impact the minima position. It is thus useless at optimization but required to evaluate the minimal $J$ value.

ModelPredictiveControl.init_stochpred — Function

init_stochpred(estim::StateEstimator, Hp)

Init the stochastic prediction matrix Ks from estim estimator for predictive control.

$\mathbf{K_s}$ is the prediction matrix of the stochastic model (composed exclusively of integrators):

\[ \mathbf{Ŷ_s} = \mathbf{K_s x̂_s}(k)\]

The stochastic predictions $\mathbf{Ŷ_s}$ are the integrator outputs from $k+1$ to $k+H_p$. $\mathbf{x̂_s}(k)$ is extracted from current estimates $\mathbf{x̂}_{k-1}(k)$ with split_state. The method also returns an empty $\mathbf{P_s}$ matrix, since it is useless except for InternalModel estimators.

init_stochpred(estim::InternalModel, Hp)

Init the stochastic prediction matrices for InternalModel.

Ks and Ps matrices are defined as:

\[ \mathbf{Ŷ_s} = \mathbf{K_s x̂_s}(k) + \mathbf{P_s ŷ_s}(k)\]

Current stochastic outputs $\mathbf{ŷ_s}(k)$ comprises the measured outputs $\mathbf{ŷ_s^m}(k) = \mathbf{y^m}(k) - \mathbf{ŷ_d^m}(k)$ and unmeasured $\mathbf{ŷ_s^u(k) = 0}$. See [2].

ModelPredictiveControl.split_state — Function

split_state(estim::StateEstimator)

Split estim.x̂ vector into the deterministic x̂d and stochastic x̂s states.

split_state(estim::InternalModel)

Get the internal model deterministic estim.x̂d and stochastic estim.x̂s states.

ModelPredictiveControl.init_prediction — Function

init_prediction(mpc, model::LinModel, d, D̂, Ŷs, R̂y, x̂d, lastu)

Init linear model prediction matrices F, q̃ and p.

See init_deterpred and init_quadprog for the definition of the matrices.

ModelPredictiveControl.init_constraint — Function

init_constraint(mpc, ::LinModel, F)

Init b vector for the linear model inequality constraints ($\mathbf{A ΔŨ ≤ b}$).

1Maciejowski, J. 2000, "Predictive control : with constraints", 1st ed., Prentice Hall, ISBN 978-0201398236.

2Desbiens, A., D. Hodouin & É. Plamondon. 2000, "Global predictive control : a unified control structure for decoupling setpoint tracking, feedforward compensation and disturbance rejection dynamics", IEE Proceedings - Control Theory and Applications, vol. 147, no 4, https://doi.org/10.1049/ip-cta:20000443, p. 465–475, ISSN 1350-2379.