Functions: Generic Functions

Functions: Generic Functions

Evaluate Output y

Update State x

Init State x

Set State x

Set Constraint

Quick Simulation

This page contains the documentation of functions that are generic to SimModel, StateEstimator and PredictiveController types.

Evaluate Output y

ModelPredictiveControl.evaloutput — Function

evaloutput(model::SimModel, d=[]) -> y

Evaluate SimModel outputs y from model.x states and measured disturbances d.

Calling a SimModel object calls this evaloutput method.

Examples

julia> model = setop!(LinModel(tf(2, [10, 1]), 5.0), yop=[20]);

julia> y = evaloutput(model)
1-element Vector{Float64}:
 20.0

evaloutput(estim::StateEstimator, d=[]) -> ŷ

Evaluate StateEstimator outputs ŷ from estim.x̂ states and disturbances d.

Calling a StateEstimator object calls this evaloutput method.

Examples

julia> kf = SteadyKalmanFilter(setop!(LinModel(tf(2, [10, 1]), 5), yop=[20]));

julia> ŷ = evaloutput(kf)
1-element Vector{Float64}:
 20.0

Update State x

ModelPredictiveControl.updatestate! — Function

updatestate!(model::SimModel, u, d=[]) -> x

Update model.x states with current inputs u and measured disturbances d.

Examples

julia> model = LinModel(ss(1.0, 1.0, 1.0, 0, 1.0));

julia> x = updatestate!(model, [1])
1-element Vector{Float64}:
 1.0

updatestate!(estim::StateEstimator, u, ym, d=[]) -> x̂

Update estim.x̂ estimate with current inputs u, measured outputs ym and dist. d.

The method removes the operating points with remove_op! and call update_estimate!.

Examples

julia> kf = SteadyKalmanFilter(LinModel(ss(0.1, 0.5, 1, 0, 4.0)));

julia> x̂ = updatestate!(kf, [1], [0]) # x̂[2] is the integrator state (nint_ym argument)
2-element Vector{Float64}:
 0.5
 0.0

updatestate!(mpc::PredictiveController, u, ym, d=[]) -> x̂

Call updatestate! on mpc.estim StateEstimator.

Init State x

ModelPredictiveControl.initstate! — Function

initstate!(model::SimModel, u, d=[]) -> x

Init model.x with manipulated inputs u and measured disturbances d steady-state.

It calls steadystate!(model, u, d):

If model is a LinModel, the method computes the steady-state of current inputs u and measured disturbances d.

Else, model.x is left unchanged. Use setstate! to manually modify it.

Examples

julia> model = LinModel(tf(6, [10, 1]), 2.0);

julia> u = [1]; x = initstate!(model, u); y = round.(evaloutput(model), digits=3)
1-element Vector{Float64}:
 6.0
 
julia> x ≈ updatestate!(model, u)
true

initstate!(estim::StateEstimator, u, ym, d=[]) -> x̂

Init estim.x̂ states from current inputs u, measured outputs ym and disturbances d.

The method tries to find a good stead-state for the initial esitmate $\mathbf{x̂}(0)$. It removes the operating points with remove_op! and call init_estimate!:

If estim.model is a LinModel, it finds the steady-state of the augmented model using u and d arguments, and uses the ym argument to enforce that $\mathbf{ŷ^m}(0) = \mathbf{y^m}(0)$. For control applications, this solution produces a bumpless manual to automatic transfer. See init_estimate! for details.

Else, estim.x̂ is left unchanged. Use setstate! to manually modify it.

If applicable, it also sets the error covariance estim.P̂ to estim.P̂0.

Examples

julia> estim = SteadyKalmanFilter(LinModel(tf(3, [10, 1]), 0.5), nint_ym=[2]);

julia> u = [1]; y = [3 - 0.1]; x̂ = round.(initstate!(estim, u, y), digits=3)
3-element Vector{Float64}:
  5.0
  0.0
 -0.1

julia> x̂ ≈ updatestate!(estim, u, y)
true

julia> evaloutput(estim) ≈ y
true

initstate!(mpc::PredictiveController, u, ym, d=[]) -> x̂

Init the states of mpc.estim StateEstimator and warm start mpc.ΔŨ at zero.

Set State x

ModelPredictiveControl.setstate! — Function

setstate!(model::SimModel, x)

Set model.x states to values specified by x.

setstate!(estim::StateEstimator, x̂)

Set estim.x̂ states to values specified by x̂.

setstate!(mpc::PredictiveController, x̂)

Set the estimate at mpc.estim.x̂.

Set Constraint

ModelPredictiveControl.setconstraint! — Function

setconstraint!(estim::MovingHorizonEstimator; <keyword arguments>) -> estim

Set the constraint parameters of estim MovingHorizonEstimator.

The constraints of the moving horizon estimator are:

\[\mathbf{x̂_{min}} ≤ \mathbf{x̂}_k(k-j+1) ≤ \mathbf{x̂_{max}} \qquad j = 0, 1, ... , N_k \\\]

Note that state constraints are applied on the augmented state vector $\mathbf{x̂}$ (see the extended help of SteadyKalmanFilter for details on augmentation).

Arguments

Info

The default constraints are mentioned here for clarity but omitting a keyword argument will not re-assign to its default value (defaults are set at construction only).

estim::MovingHorizonEstimator : moving horizon estimator to set constraints.

x̂min = fill(-Inf,nx̂) : augmented state vector lower bounds $\mathbf{x̂_{min}}$.

x̂max = fill(+Inf,nx̂) : augmented state vector upper bounds $\mathbf{x̂_{max}}$.

all the keyword arguments above but with a capital letter, e.g. X̂max or X̂min : for time-varying constraints (see Extended Help).

Examples

julia> estim = MovingHorizonEstimator(LinModel(ss(0.5,1,1,0,1)), He=3);

julia> estim = setconstraint!(estim, x̂min=[-50, -50], x̂max=[50, 50])
MovingHorizonEstimator estimator with a sample time Ts = 1.0 s, LinModel and:
 3 estimation steps He
 1 manipulated inputs u (0 integrating states)
 2 states x̂
 1 measured outputs ym (1 integrating states)
 0 unmeasured outputs yu
 0 measured disturbances d

setconstraint!(mpc::PredictiveController; <keyword arguments>) -> mpc

Set the constraint parameters of mpc predictive controller.

The predictive controllers support both soft and hard constraints, defined by:

\[\begin{alignat*}{3} \mathbf{u_{min} - c_{u_{min}}} ϵ ≤&&\ \mathbf{u}(k+j) &≤ \mathbf{u_{max} + c_{u_{max}}} ϵ &&\qquad j = 0, 1 ,..., H_p - 1 \\ \mathbf{Δu_{min} - c_{Δu_{min}}} ϵ ≤&&\ \mathbf{Δu}(k+j) &≤ \mathbf{Δu_{max} + c_{Δu_{max}}} ϵ &&\qquad j = 0, 1 ,..., H_c - 1 \\ \mathbf{y_{min} - c_{y_{min}}} ϵ ≤&&\ \mathbf{ŷ}(k+j) &≤ \mathbf{y_{max} + c_{y_{max}}} ϵ &&\qquad j = 1, 2 ,..., H_p \\ \mathbf{x̂_{min} - c_{x̂_{min}}} ϵ ≤&&\ \mathbf{x̂}_{k-1}(k+j) &≤ \mathbf{x̂_{max} + c_{x̂_{max}}} ϵ &&\qquad j = H_p \end{alignat*}\]

and also $ϵ ≥ 0$. The last line is the terminal constraints applied on the states at the end of the horizon (see Extended Help). All the constraint parameters are vector. Use ±Inf values when there is no bound. The constraint softness parameters $\mathbf{c}$, also called equal concern for relaxation, are non-negative values that specify the softness of the associated bound. Use 0.0 values for hard constraints. The output and terminal constraints are all soft by default. See Extended Help for time-varying constraints.

Arguments

Info

Same as above for the keyword arguments and default constraints.

mpc::PredictiveController : predictive controller to set constraints.

umin = fill(-Inf,nu) : manipulated input lower bounds $\mathbf{u_{min}}$.

umax = fill(+Inf,nu) : manipulated input upper bounds $\mathbf{u_{max}}$.

Δumin = fill(-Inf,nu) : manipulated input increment lower bounds $\mathbf{Δu_{min}}$.

Δumax = fill(+Inf,nu) : manipulated input increment upper bounds $\mathbf{Δu_{max}}$.

ymin = fill(-Inf,ny) : predicted output lower bounds $\mathbf{y_{min}}$.

ymax = fill(+Inf,ny) : predicted output upper bounds $\mathbf{y_{max}}$.

x̂min = fill(-Inf,nx̂) : terminal constraint lower bounds $\mathbf{x̂_{min}}$.

x̂max = fill(+Inf,nx̂) : terminal constraint upper bounds $\mathbf{x̂_{max}}$.

c_umin = fill(0.0,nu) : umin softness weights $\mathbf{c_{u_{min}}}$.

c_umax = fill(0.0,nu) : umax softness weights $\mathbf{c_{u_{max}}}$.

c_Δumin = fill(0.0,nu) : Δumin softness weights $\mathbf{c_{Δu_{min}}}$.

c_Δumax = fill(0.0,nu) : Δumax softness weights $\mathbf{c_{Δu_{max}}}$.

c_ymin = fill(1.0,ny) : ymin softness weights $\mathbf{c_{y_{min}}}$.

c_ymax = fill(1.0,ny) : ymax softness weights $\mathbf{c_{y_{max}}}$.

c_x̂min = fill(1.0,nx̂) : x̂min softness weights $\mathbf{c_{x̂_{min}}}$.

c_x̂max = fill(1.0,nx̂) : x̂max softness weights $\mathbf{c_{x̂_{max}}}$.

all the keyword arguments above but with a capital letter, except for the terminal constraints, e.g. Ymax or C_Δumin : for time-varying constraints (see Extended Help).

Examples

julia> mpc = LinMPC(setop!(LinModel(tf(3, [30, 1]), 4), uop=[50], yop=[25]));

julia> mpc = setconstraint!(mpc, umin=[0], umax=[100], c_umin=[0.0], c_umax=[0.0]);

julia> mpc = setconstraint!(mpc, Δumin=[-10], Δumax=[+10], c_Δumin=[1.0], c_Δumax=[1.0])
LinMPC controller with a sample time Ts = 4.0 s, OSQP optimizer, SteadyKalmanFilter estimator and:
 10 prediction steps Hp
  2 control steps Hc
  1 manipulated inputs u (0 integrating states)
  2 states x̂
  1 measured outputs ym (1 integrating states)
  0 unmeasured outputs yu
  0 measured disturbances d

Extended Help

Terminal constraints provide closed-loop stability guarantees on the nominal plant model. They can render an unfeasible problem however. In practice, a sufficiently large prediction horizon $H_p$ without terminal constraints is typically enough for stability. Note that terminal constraints are applied on the augmented state vector $\mathbf{x̂}$ (see SteadyKalmanFilter for details on augmentation).

For variable constraints, the bounds can be modified after calling moveinput!, that is, at runtime, but not the softness parameters $\mathbf{c}$. It is not possible to modify ±Inf bounds at runtime.

Tip

To keep a variable unconstrained while maintaining the ability to add a constraint later at runtime, set the bound to an absolute value sufficiently large when you create the controller (but different than ±Inf).

It is also possible to specify time-varying constraints over $H_p$ and $H_c$ horizons. In such a case, they are defined by:

\[\begin{alignat*}{3} \mathbf{U_{min} - C_{u_{min}}} ϵ ≤&&\ \mathbf{U} &≤ \mathbf{U_{max} + C_{u_{max}}} ϵ \\ \mathbf{ΔU_{min} - C_{Δu_{min}}} ϵ ≤&&\ \mathbf{ΔU} &≤ \mathbf{ΔU_{max} + C_{Δu_{max}}} ϵ \\ \mathbf{Y_{min} - C_{y_{min}}} ϵ ≤&&\ \mathbf{Ŷ} &≤ \mathbf{Y_{max} + C_{y_{max}}} ϵ \end{alignat*}\]

For this, use the same keyword arguments as above but with a capital letter:

Umin / Umax / C_umin / C_umax : $\mathbf{U}$ constraints (nu*Hp,).

ΔUmin / ΔUmax / C_Δumin / C_Δumax : $\mathbf{ΔU}$ constraints (nu*Hc,).

Ymin / Ymax / C_ymin / C_ymax : $\mathbf{Ŷ}$ constraints (ny*Hp,).

Quick Simulation

Simulate

ModelPredictiveControl.sim! — Function

sim!(plant::SimModel, N::Int, u=plant.uop.+1, d=plant.dop; x0=zeros(plant.nx)) -> res

Open-loop simulation of plant for N time steps, default to unit bump test on all inputs.

The manipulated inputs $\mathbf{u}$ and measured disturbances $\mathbf{d}$ are held constant at u and d values, respectively. The plant initial state $\mathbf{x}(0)$ is specified by x0 keyword arguments. The function returns SimResult instances that can be visualized by calling plot from Plots.jl on them (see Examples below). Note that the method mutates plant internal states.

Examples

julia> plant = NonLinModel((x,u,d)->0.1x+u+d, (x,_)->2x, 10.0, 1, 1, 1, 1);

julia> res = sim!(plant, 15, [0], [0], x0=[1])
Simulation results of NonLinModel with 15 time steps.

julia> using Plots; plot(res, plotu=false, plotd=false, plotx=true)

sim!(
    estim::StateEstimator,
    N::Int,
    u = estim.model.uop .+ 1,
    d = estim.model.dop;
    <keyword arguments>
) -> res

Closed-loop simulation of estim estimator for N steps, default to input bumps.

See Arguments for the available options. The noises are provided as standard deviations σ vectors. The simulated sensor and process noises of plant are specified by y_noise and x_noise arguments, respectively.

Arguments

estim::StateEstimator : state estimator to simulate.

N::Int : simulation length in time steps.

u = estim.model.uop .+ 1 : manipulated input $\mathbf{u}$ value.

d = estim.model.dop : plant measured disturbance $\mathbf{d}$ value.

plant::SimModel = estim.model : simulated plant model.

u_step = zeros(plant.nu) : step load disturbance on plant inputs $\mathbf{u}$.

u_noise = zeros(plant.nu) : gaussian load disturbance on plant inputs $\mathbf{u}$.

y_step = zeros(plant.ny) : step disturbance on plant outputs $\mathbf{y}$.

y_noise = zeros(plant.ny) : additive gaussian noise on plant outputs $\mathbf{y}$.

d_step = zeros(plant.nd) : step on measured disturbances $\mathbf{d}$.

d_noise = zeros(plant.nd) : additive gaussian noise on measured dist. $\mathbf{d}$.

x_noise = zeros(plant.nx) : additive gaussian noise on plant states $\mathbf{x}$.

x0 = plant.x : plant initial state $\mathbf{x}(0)$.

x̂0 = nothing : initial estimate $\mathbf{x̂}(0)$, initstate! is used if nothing.

lastu = plant.uop : last plant input $\mathbf{u}$ for $\mathbf{x̂}$ initialization.

Examples

julia> model = LinModel(tf(3, [30, 1]), 0.5);

julia> estim = KalmanFilter(model, σR=[0.5], σQ=[0.25], σQint_ym=[0.01], σP0int_ym=[0.1]);

julia> res = sim!(estim, 50, [0], y_noise=[0.5], x_noise=[0.25], x0=[-10], x̂0=[0, 0])
Simulation results of KalmanFilter with 50 time steps.

julia> using Plots; plot(res, plotŷ=true, plotu=false, plotxwithx̂=true)

sim!(
    mpc::PredictiveController, 
    N::Int,
    ry = mpc.estim.model.yop .+ 1, 
    d  = mpc.estim.model.dop,
    ru = mpc.estim.model.uop;
    <keyword arguments>
) -> res

Closed-loop simulation of mpc controller for N steps, default to output setpoint bumps.

The output and manipulated input setpoints are held constant at ry and ru, respectively. The keyword arguments are identical to sim!(::StateEstimator, ::Int).

Examples

julia> model = LinModel([tf(3, [30, 1]); tf(2, [5, 1])], 4);

julia> mpc = setconstraint!(LinMPC(model, Mwt=[0, 1], Nwt=[0.01], Hp=30), ymin=[0, -Inf]);

julia> res = sim!(mpc, 25, [0, 0], y_noise=[0.1], y_step=[-10, 0])
Simulation results of LinMPC with 25 time steps.

julia> using Plots; plot(res, plotry=true, plotŷ=true, plotymin=true, plotu=true)

Simulation Results

ModelPredictiveControl.SimResult — Type

SimResult(
    obj::Union{SimModel, StateEstimator, PredictiveController}, 
    U_data, 
    Y_data, 
    D_data  = [];
    X_data  = nothing,
    X̂_data  = nothing, 
    Ry_data = nothing, 
    Ru_data = nothing
)

Manually construct a SimResult to quickly plot obj simulations.

Except for obj, all the arguments should be matrices of N columns, where N is the number of time steps. SimResult objects allow to quickly plot simulation results. Simply call plot from Plots.jl on them.

Examples

julia> plant = LinModel(tf(1, [1, 1]), 1.0); N = 5; U_data = fill(1.0, 1, N);

julia> Y_data = reduce(hcat, (updatestate!(plant, U_data[:, i]); plant()) for i=1:N)
1×5 Matrix{Float64}:
 0.632121  0.864665  0.950213  0.981684  0.993262

julia> res = SimResult(plant, U_data, Y_data)
Simulation results of LinModel with 5 time steps.

julia> using Plots; plot(res)