Functions: Generic Functions

Functions: Generic Functions

Evaluate Output y

Update State x

Init State x

Set State x

Quick Simulation

This page contains the documentation of functions that are generic to SimModel, StateEstimator and PredictiveController types.

Evaluate Output y

ModelPredictiveControl.evaloutput — Function

evaloutput(model::SimModel, d=[]) -> y

Evaluate SimModel outputs y from model.x states and measured disturbances d.

Calling a SimModel object calls this evaloutput method.

Examples

julia> model = setop!(LinModel(tf(2, [10, 1]), 5.0), yop=[20]);

julia> y = evaloutput(model)
1-element Vector{Float64}:
 20.0

evaloutput(estim::StateEstimator, d=[]) -> ŷ

Evaluate StateEstimator outputs ŷ from estim.x̂ states and disturbances d.

Calling a StateEstimator object calls this evaloutput method.

Examples

julia> kf = SteadyKalmanFilter(setop!(LinModel(tf(2, [10, 1]), 5), yop=[20]));

julia> ŷ = evaloutput(kf)
1-element Vector{Float64}:
 20.0

Update State x

ModelPredictiveControl.updatestate! — Function

updatestate!(model::SimModel, u, d=[]) -> x

Update model.x states with current inputs u and measured disturbances d.

Examples

julia> model = LinModel(ss(1, 1, 1, 0, 1.0));

julia> x = updatestate!(model, [1])
1-element Vector{Float64}:
 1.0

updatestate!(estim::StateEstimator, u, ym, d=[]) -> x̂

Update estim.x̂ estimate with current inputs u, measured outputs ym and dist. d.

The method removes the operating points with remove_op! and call update_estimate!.

Examples

julia> kf = SteadyKalmanFilter(LinModel(ss(0.1, 0.5, 1, 0, 4.0)));

julia> x̂ = updatestate!(kf, [1], [0]) # x̂[2] is the integrator state (nint_ym argument)
2-element Vector{Float64}:
 0.5
 0.0

updatestate!(mpc::PredictiveController, u, ym, d=[]) -> x̂

Call updatestate! on mpc.estim StateEstimator.

Init State x

ModelPredictiveControl.initstate! — Function

initstate!(model::SimModel, u, d=[]) -> x

Init model.x with manipulated inputs u and measured disturbances d steady-state.

It calls steadystate!(model, u, d):

If model is a LinModel, the method computes the steady-state of current inputs u and measured disturbances d.

Else, model.x is left unchanged. Use setstate! to manually modify it.

Examples

julia> model = LinModel(tf(6, [10, 1]), 2.0);

julia> u = [1]; x = initstate!(model, u); y = round.(evaloutput(model), digits=3)
1-element Vector{Float64}:
 6.0
 
julia> x ≈ updatestate!(model, u)
true

initstate!(estim::StateEstimator, u, ym, d=[]) -> x̂

Init estim.x̂ states from current inputs u, measured outputs ym and disturbances d.

The method tries to find a good stead-state for the initial esitmate $\mathbf{x̂}(0)$. It removes the operating points with remove_op! and call init_estimate!:

If estim.model is a LinModel, it finds the steady-state of the augmented model using u and d arguments, and uses the ym argument to enforce that $\mathbf{ŷ^m}(0) = \mathbf{y^m}(0)$. For control applications, this solution produces a bumpless manual to automatic transfer. See init_estimate! for details.

Else, estim.x̂ is left unchanged. Use setstate! to manually modify it.

If applicable, it also sets the error covariance estim.P̂ to estim.P̂0.

Examples

julia> estim = SteadyKalmanFilter(LinModel(tf(3, [10, 1]), 0.5), nint_ym=[2]);

julia> u = [1]; y = [3 - 0.1]; x̂ = round.(initstate!(estim, u, y), digits=3)
3-element Vector{Float64}:
  5.0
  0.0
 -0.1

julia> x̂ ≈ updatestate!(estim, u, y)
true

julia> evaloutput(estim) ≈ y
true

initstate!(mpc::PredictiveController, u, ym, d=[]) -> x̂

Init the states of mpc.estim StateEstimator and warm start mpc.ΔŨ at zero.

Set State x

ModelPredictiveControl.setstate! — Function

setstate!(model::SimModel, x)

Set model.x states to values specified by x.

setstate!(estim::StateEstimator, x̂)

Set estim.x̂ states to values specified by x̂.

setstate!(mpc::PredictiveController, x̂)

Set the estimate at mpc.estim.x̂.

Quick Simulation

ModelPredictiveControl.sim! — Function

sim!(plant::SimModel, N::Int, u=plant.uop.+1, d=plant.dop; x0=zeros(plant.nx)) -> res

Open-loop simulation of plant for N time steps, default to unit bump test on all inputs.

The manipulated inputs $\mathbf{u}$ and measured disturbances $\mathbf{d}$ are held constant at u and d values, respectively. The plant initial state $\mathbf{x}(0)$ is specified by x0 keyword arguments. The function returns SimResult instances that can be visualized by calling plot from Plots.jl on them (see Examples below). Note that the method mutates plant internal states.

Examples

julia> plant = NonLinModel((x,u,d)->0.1x+u+d, (x,_)->2x, 10.0, 1, 1, 1, 1);

julia> res = sim!(plant, 15, [0], [0], x0=[1]);

julia> using Plots; plot(res, plotu=false, plotd=false, plotx=true)

sim!(
    estim::StateEstimator,
    N::Int,
    u = estim.model.uop .+ 1,
    d = estim.model.dop;
    <keyword arguments>
) -> res

Closed-loop simulation of estim estimator for N steps, default to input bumps.

See Arguments for the available options. The noises are provided as standard deviations σ vectors. The simulated sensor and process noises of plant are specified by y_noise and x_noise arguments, respectively.

Arguments

estim::StateEstimator : state estimator to simulate.

N::Int : simulation length in time steps.

u = estim.model.uop .+ 1 : manipulated input $\mathbf{u}$ value.

d = estim.model.dop : plant measured disturbance $\mathbf{d}$ value.

plant::SimModel = estim.model : simulated plant model.

u_step = zeros(plant.nu) : step load disturbance on plant inputs $\mathbf{u}$.

u_noise = zeros(plant.nu) : gaussian load disturbance on plant inputs $\mathbf{u}$.

y_step = zeros(plant.ny) : step disturbance on plant outputs $\mathbf{y}$.

y_noise = zeros(plant.ny) : additive gaussian noise on plant outputs $\mathbf{y}$.

d_step = zeros(plant.nd) : step on measured disturbances $\mathbf{d}$.

d_noise = zeros(plant.nd) : additive gaussian noise on measured dist. $\mathbf{d}$.

x_noise = zeros(plant.nx) : additive gaussian noise on plant states $\mathbf{x}$.

x0 = zeros(plant.nx) : plant initial state $\mathbf{x}(0)$.

x̂0 = nothing : initial estimate $\mathbf{x̂}(0)$, initstate! is used if nothing.

lastu = plant.uop : last plant input $\mathbf{u}$ for $\mathbf{x̂}$ initialization.

Examples

julia> model = LinModel(tf(3, [30, 1]), 0.5);

julia> estim = KalmanFilter(model, σR=[0.5], σQ=[0.25], σQ_int=[0.01], σP0_int=[0.1]);

julia> res = sim!(estim, 50, [0], y_noise=[0.5], x_noise=[0.25], x0=[-10], x̂0=[0, 0]);

julia> using Plots; plot(res, plotŷ=true, plotu=false, plotxwithx̂=true)

sim!(
    mpc::PredictiveController, 
    N::Int,
    ry = mpc.estim.model.yop .+ 1, 
    d  = mpc.estim.model.dop,
    ru = mpc.estim.model.uop;
    <keyword arguments>
) -> res

Closed-loop simulation of mpc controller for N steps, default to output setpoint bumps.

The output and manipulated input setpoints are held constant at ry and ru, respectively. The keyword arguments are identical to sim!(::StateEstimator, ::Int).

Examples

julia> model = LinModel([tf(3, [30, 1]); tf(2, [5, 1])], 4);

julia> mpc = setconstraint!(LinMPC(model, Mwt=[0, 1], Nwt=[0.01], Hp=30), ymin=[0, -Inf]);

julia> res = sim!(mpc, 25, [0, 0], y_noise=[0.1], y_step=[-10, 0]);

julia> using Plots; plot(res, plotry=true, plotŷ=true, plotymin=true, plotu=true)