Generic Functions

Generic Functions

Evaluate Output y

Update State x

Init State x

Set State x

Quick Simulation

Evaluate Output y

ModelPredictiveControl.evaloutput — Function

evaloutput(model::SimModel, d=Float64[])

Evaluate SimModel outputs y from model.x states and measured disturbances d.

Calling a SimModel object calls this evaloutput method.

Examples

julia> model = setop!(LinModel(tf(2, [10, 1]), 5), yop=[20]);

julia> y = evaloutput(model)
1-element Vector{Float64}:
 20.0

evaloutput(estim::StateEstimator, d=Float64[])

Evaluate StateEstimator outputs ŷ from estim.x̂ states and disturbances d.

Calling a StateEstimator object calls this evaloutput method.

Examples

julia> kf = SteadyKalmanFilter(setop!(LinModel(tf(2, [10, 1]), 5), yop=[20]));

julia> ŷ = evaloutput(kf)
1-element Vector{Float64}:
 20.0

evaloutput(estim::InternalModel, ym, d=Float64[])

Evaluate InternalModel outputs ŷ from estim.x̂d states and measured outputs ym.

InternalModel estimator needs current measured outputs $\mathbf{y^m}(k)$ to estimate its outputs $\mathbf{ŷ}(k)$, since the strategy imposes that $\mathbf{ŷ^m}(k) = \mathbf{y^m}(k)$ is always true.

Update State x

ModelPredictiveControl.updatestate! — Function

updatestate!(model::SimModel, u, d=Float64[])

Update model.x states with current inputs u and measured disturbances d.

Examples

julia> model = LinModel(ss(1, 1, 1, 0, 1));

julia> x = updatestate!(model, [1])
1-element Vector{Float64}:
 1.0

updatestate!(estim::SteadyKalmanFilter, u, ym, d=Float64[])

Update estim.x̂ estimate with current inputs u, measured outputs ym and dist. d.

The SteadyKalmanFilter updates it with the precomputed Kalman gain $\mathbf{K}$:

\[\mathbf{x̂}_{k}(k+1) = \mathbf{Â x̂}_{k-1}(k) + \mathbf{B̂_u u}(k) + \mathbf{B̂_d d}(k) + \mathbf{K}[\mathbf{y^m}(k) - \mathbf{Ĉ^m x̂}_{k-1}(k) - \mathbf{D̂_d^m d}(k)]\]

Examples

julia> kf = SteadyKalmanFilter(LinModel(ss(1, 1, 1, 0, 1)));

julia> x̂ = updatestate!(kf, [1], [0]) # x̂[2] is the integrator state (nint_ym argument)
2-element Vector{Float64}:
 1.0
 0.0

updatestate!(estim::KalmanFilter, u, ym, d=Float64[])

Update estim.x̂ \ P̂ with current inputs u, measured outputs ym and dist. d.

See updatestate_kf! for the implementation details.

updatestate!(estim::UnscentedKalmanFilter, u, ym, d=Float64[])

Same than KalmanFilter but using the unscented estimator.

See updatestate_ukf! for the implementation details.

updatestate!(estim::InternalModel, u, ym, d=Float64[])

Update estim.x̂ \ x̂d \ x̂s with current inputs u, measured outputs ym and dist. d.

updatestate!(mpc::PredictiveController, u, ym, d=Float64[])

Call updatestate! on mpc.estim StateEstimator.

Init State x

ModelPredictiveControl.initstate! — Function

initstate!(estim::StateEstimator, u, ym, d=Float64[])

Init estim.x̂ states from current inputs u, measured outputs ym and disturbances d.

The method tries to find a good steady-state to initialize estim.x̂ estimate :

If estim.model is a LinModel, it evaluates estim.model steady-state (using steadystate) with current inputs u and measured disturbances d, and saves the result to estim.x̂[1:nx].

Else, the current deterministic states estim.x̂[1:nx] are left unchanged (use setstate! to manually modify them).

It then estimates the measured outputs ŷm from these states, and the residual offset with current measured outputs (ym - ŷm) initializes the integrators of the stochastic model. This approach ensures that $\mathbf{ŷ^m}(0) = \mathbf{y^m}(0)$. For LinModel, it also ensures that the estimator starts at steady-state, resulting in a bumpless manual to automatic transfer for control applications.

Examples

julia> estim = SteadyKalmanFilter(LinModel(tf(3, [10, 1]), 0.5), nint_ym=[2]);

julia> x̂ = initstate!(estim, [1], [3 - 0.1])
3-element Vector{Float64}:
  5.0000000000000115
  0.0
 -0.10000000000000675

initstate!(estim::Union{KalmanFilter, UnscentedKalmanFilter}, u, ym, d=Float64[])

Initialize covariance estim.P̂ and invoke initstate!(::StateEstimator).

initstate!(estim::InternalModel, u, ym, d=Float64[])

Init estim.x̂d / x̂s states from current inputs u, meas. outputs ym and disturb. d.

The deterministic state estim.x̂d initialization method is identical to initstate!(::StateEstimator). The stochastic states estim.x̂s are init at 0.

initstate!(mpc::PredictiveController, u, ym, d=Float64[])

Init mpc.ΔŨ for warm-starting and the states of mpc.estim StateEstimator.

Set State x

ModelPredictiveControl.setstate! — Function

setstate!(model::SimModel, x)

Set model.x states to values specified by x.

setstate!(estim::StateEstimator, x̂)

Set estim.x̂ states to values specified by x̂.

setstate!(mpc::PredictiveController, x̂)

Set the estimate at mpc.estim.x̂.

Quick Simulation

ModelPredictiveControl.sim — Function

sim(
    mpc::PredictiveController, 
    N::Int, 
    ry = mpc.estim.model.yop .+ 1, 
    d  = mpc.estim.model.dop; 
    <keyword arguments>
)

Closed-loop simulation of mpc controller for N time steps, default to setpoint bumps.

See Arguments for the option list. The noise arguments are in standard deviations σ. The sensor and process noises of the simulated plant are specified by y_noise and x_noise arguments, respectively.

Arguments

mpc::PredictiveController : predictive controller to simulate

N::Int : simulation length in time steps

ry = mpc.estim.model.yop .+ 1 : plant output setpoint $\mathbf{r_y}$ value

d = mpc.estim.model.dop : plant measured disturbance $\mathbf{d}$ value

plant::SimModel = mpc.estim.model : simulated plant model

u_step = zeros(plant.nu) : step disturbance on manipulated input $\mathbf{u}$

u_noise = zeros(plant.nu) : additive gaussian noise on manipulated input $\mathbf{u}$

y_step = zeros(plant.ny) : step disturbance on plant outputs $\mathbf{y}$

y_noise = zeros(plant.ny) : additive gaussian noise on plant outputs $\mathbf{y}$

d_step = zeros(plant.nd) : step disturbance on measured dist. $\mathbf{d}$

d_noise = zeros(plant.nd) : additive gaussian noise on measured dist. $\mathbf{d}$

x_noise = zeros(plant.nx) : additive gaussian noise on plant states $\mathbf{x}$

x0 = zeros(plant.nx) : plant initial state $\mathbf{x}(0)$

x̂0 = nothing : mpc.estim state estimator initial state $\mathbf{x̂}(0)$, if nothing then $\mathbf{x̂}$ is initialized with initstate!

lastu = plant.uop : last plant input $\mathbf{u}$ for $\mathbf{x̂}$ initialization