Predictive Controllers

Predictive Controllers

PredictiveController

LinMPC

NonLinMPC

Set Constraint

Move Manipulated Input

Get Additional Information

All the predictive controllers in this module rely on a state estimator to compute the predictions. The default LinMPC estimator is a SteadyKalmanFilter, and NonLinMPC, an UnscentedKalmanFilter. For simpler and more classical designs, an InternalModel structure is also available, that assumes by default that the current model mismatch estimation is constant in the future (same approach than dynamic matrix control, DMC).

Info

The nomenclature uses hats for the predictions (or estimations, for the state estimators) e.g. $\mathbf{Ŷ}$ encompasses the future model outputs $\mathbf{y}$ over the prediction horizon $H_p$.

PredictiveController

ModelPredictiveControl.PredictiveController — Type

Abstract supertype of all predictive controllers.

(mpc::PredictiveController)(ry, d=Float64[]; kwargs...)

Functor allowing callable PredictiveController object as an alias for moveinput!.

Examples

julia> mpc = LinMPC(LinModel(tf(5, [2, 1]), 3), Nwt=[0], Hp=1000, Hc=1);

julia> u = mpc([5]); round.(u, digits=3)
1-element Vector{Float64}:
 1.0

LinMPC

ModelPredictiveControl.LinMPC — Type

LinMPC(model::LinModel; <keyword arguments>)

Construct a linear predictive controller based on LinModel model.

The controller minimizes the following objective function at each discrete time $k$:

\[\min_{\mathbf{ΔU}, ϵ} \mathbf{(R̂_y - Ŷ)}' \mathbf{M}_{H_p} \mathbf{(R̂_y - Ŷ)} + \mathbf{(ΔU)}' \mathbf{N}_{H_c} \mathbf{(ΔU)} + \mathbf{(R̂_u - U)}' \mathbf{L}_{H_p} \mathbf{(R̂_u - U)} + C ϵ^2\]

in which the weight matrices are repeated $H_p$ or $H_c$ times:

\[\begin{aligned} \mathbf{M}_{H_p} &= \text{diag}\mathbf{(M,M,...,M)} \\ \mathbf{N}_{H_c} &= \text{diag}\mathbf{(N,N,...,N)} \\ \mathbf{L}_{H_p} &= \text{diag}\mathbf{(L,L,...,L)} \end{aligned}\]

and with the following nomenclature:

VARIABLE	DESCRIPTION	SIZE

$H_p$	prediction horizon (integer)	()

$H_c$	control horizon (integer)	()

$\mathbf{ΔU}$	manipulated input increments over $H_c$	(nu*Hc,)

$\mathbf{Ŷ}$	predicted outputs over $H_p$	(ny*Hp,)

$\mathbf{U}$	manipulated inputs over $H_p$	(nu*Hp,)

$\mathbf{R̂_y}$	predicted output setpoints over $H_p$	(ny*Hp,)

$\mathbf{R̂_u}$	predicted manipulated input setpoints over $H_p$	(nu*Hp,)

$\mathbf{M}$	output setpoint tracking weights	(ny*Hp, ny*Hp)

$\mathbf{N}$	manipulated input increment weights	(nu*Hc, nu*Hc)

$\mathbf{L}$	manipulated input setpoint tracking weights	(nu*Hp, nu*Hp)

$C$	slack variable weight	()

$ϵ$	slack variable for constraint softening	()

The $\mathbf{ΔU}$ vector includes the manipulated input increments $\mathbf{Δu}(k+j) = \mathbf{u}(k+j) - \mathbf{u}(k+j-1)$ from $j=0$ to $H_c-1$, the $\mathbf{Ŷ}$ vector, the output predictions $\mathbf{ŷ(k+j)}$ from $j=1$ to $H_p$, and the $\mathbf{U}$ vector, the manipulated inputs $\mathbf{u}(k+j)$ from $j=0$ to $H_p-1$. The manipulated input setpoint predictions $\mathbf{R̂_u}$ are constant at $\mathbf{r_u}$.

This method uses the default state estimator, a SteadyKalmanFilter with default arguments.

Arguments

model::LinModel : model used for controller predictions and state estimations.

Hp=10+nk: prediction horizon $H_p$, nk is the number of delays in model.

Hc=2 : control horizon $H_c$.

Mwt=fill(1.0,model.ny) : main diagonal of $\mathbf{M}$ weight matrix (vector)

Nwt=fill(0.1,model.nu) : main diagonal of $\mathbf{N}$ weight matrix (vector)

Lwt=fill(0.0,model.nu) : main diagonal of $\mathbf{L}$ weight matrix (vector)

Cwt=1e5 : slack variable weight $C$ (scalar), use Cwt=Inf for hard constraints only

ru=model.uop : manipulated input setpoints $\mathbf{r_u}$ (vector)

optim=JuMP.Model(OSQP.MathOptInterfaceOSQP.Optimizer) : quadratic optimizer used in the predictive controller, provided as a JuMP.Model (default to OSQP.jl optimizer)

Examples

julia> model = LinModel([tf(3, [30, 1]); tf(-2, [5, 1])], 4);

julia> mpc = LinMPC(model, Mwt=[0, 1], Nwt=[0.5], Hp=30, Hc=1)
LinMPC controller with a sample time Ts = 4.0 s, SteadyKalmanFilter estimator and:
 1 manipulated inputs u
 4 states x̂
 2 measured outputs ym
 0 unmeasured outputs yu
 0 measured disturbances d

Extended Help

Manipulated inputs setpoints $\mathbf{r_u}$ are not common but they can be interesting for over-actuated systems, when nu > ny (e.g. prioritize solutions with lower economical costs). The default Lwt value implies that this feature is disabled by default.

LinMPC(estim::StateEstimator; <keyword arguments>)

Use custom state estimator estim to construct LinMPC.

estim.model must be a LinModel. Else, a NonLinMPC is required.

Examples

julia> estim = KalmanFilter(LinModel([tf(3, [30, 1]); tf(-2, [5, 1])], 4), i_ym=[2]);

julia> mpc = LinMPC(estim, Mwt=[0, 1], Nwt=[0.5], Hp=30, Hc=1)
LinMPC controller with a sample time Ts = 4.0 s, KalmanFilter estimator and:
 1 manipulated inputs u
 3 states x̂
 1 measured outputs ym
 1 unmeasured outputs yu
 0 measured disturbances d

NonLinMPC

ModelPredictiveControl.NonLinMPC — Type

NonLinMPC(model::SimModel; <keyword arguments>)

Construct a nonlinear predictive controller based on SimModel model.

Both NonLinModel and LinModel are supported (see Extended Help). The controller minimizes the following objective function at each discrete time $k$:

\[\min_{\mathbf{ΔU}, ϵ} \mathbf{(R̂_y - Ŷ)}' \mathbf{M}_{H_p} \mathbf{(R̂_y - Ŷ)} + \mathbf{(ΔU)}' \mathbf{N}_{H_c} \mathbf{(ΔU)} + \mathbf{(R̂_u - U)}' \mathbf{L}_{H_p} \mathbf{(R̂_u - U)} + C ϵ^2 + E J_E(\mathbf{U}_E, \mathbf{Ŷ}_E, \mathbf{D̂}_E)\]

See LinMPC for the variable definitions. The custom economic function $J_E$ can penalizes solutions with high economic costs. Setting all the weights to 0 except $E$ creates a pure economic model predictive controller (EMPC). The arguments of $J_E$ are the manipulated inputs, the predicted outputs and measured disturbances from $k$ to $k+H_p$ inclusively:

\[ \mathbf{U}_E = \begin{bmatrix} \mathbf{U} \\ \mathbf{u}(k+H_p-1) \end{bmatrix} \text{,} \qquad \mathbf{Ŷ}_E = \begin{bmatrix} \mathbf{ŷ}(k) \\ \mathbf{Ŷ} \end{bmatrix} \text{,} \qquad \mathbf{D̂}_E = \begin{bmatrix} \mathbf{d}(k) \\ \mathbf{D̂} \end{bmatrix}\]

since $H_c ≤ H_p$ implies that $\mathbf{u}(k+H_p) = \mathbf{u}(k+H_p-1)$. The vector $\mathbf{D̂}$ includes the predicted measured disturbance over $H_p$.

Tip

Replace any of the 3 arguments with _ if not needed (see JE default value below).

This method uses the default state estimator, an UnscentedKalmanFilter with default arguments.

Warning

See Extended Help if you get an error like MethodError: no method matching Float64(::ForwardDiff.Dual).

Arguments

model::SimModel : model used for controller predictions and state estimations.

Hp=10: prediction horizon $H_p$.

Hc=2 : control horizon $H_c$.

Mwt=fill(1.0,model.ny) : main diagonal of $\mathbf{M}$ weight matrix (vector)

Nwt=fill(0.1,model.nu) : main diagonal of $\mathbf{N}$ weight matrix (vector)

Lwt=fill(0.0,model.nu) : main diagonal of $\mathbf{L}$ weight matrix (vector)

Cwt=1e5 : slack variable weight $C$ (scalar), use Cwt=Inf for hard constraints only

Ewt=0.0 : economic costs weight $E$ (scalar).

JE=(_,_,_)->0.0 : economic function $J_E(\mathbf{U}_E, \mathbf{D̂}_E, \mathbf{Ŷ}_E)$.

ru=model.uop : manipulated input setpoints $\mathbf{r_u}$ (vector)

optim=JuMP.Model(Ipopt.Optimizer) : nonlinear optimizer used in the predictive controller, provided as a JuMP.Model (default to Ipopt.jl optimizer)

Examples

julia> model = NonLinModel((x,u,_)->0.5x+u, (x,_)->2x, 10, 1, 1, 1);

julia> mpc = NonLinMPC(model, Hp=20, Hc=1, Cwt=1e6)
NonLinMPC controller with a sample time Ts = 10.0 s, UnscentedKalmanFilter estimator and:
 1 manipulated inputs u
 2 states x̂
 1 measured outputs ym
 0 unmeasured outputs yu
 0 measured disturbances d

Extended Help

NonLinMPC controllers based on LinModel compute the predictions with matrix algebra instead of a for loop. This feature can accelerate the optimization and is not available in any other package, to my knowledge.

The optimizations rely on JuMP.jl automatic differentiation (AD) to compute the objective and constraint derivatives. Optimizers generally benefit from exact derivatives like AD. However, the NonLinModel f and h functions must be compatible with this feature. See Automatic differentiation for common mistakes when writing these functions.

NonLinMPC(estim::StateEstimator; <keyword arguments>)

Use custom state estimator estim to construct NonLinMPC.

Examples

julia> model = NonLinModel((x,u,_)->0.5x+u, (x,_)->2x, 10, 1, 1, 1);

julia> estim = UnscentedKalmanFilter(model, σQ_int=[0.05]);

julia> mpc = NonLinMPC(estim, Hp=20, Hc=1, Cwt=1e6)
NonLinMPC controller with a sample time Ts = 10.0 s, UnscentedKalmanFilter estimator and:
 1 manipulated inputs u
 2 states x̂
 1 measured outputs ym
 0 unmeasured outputs yu
 0 measured disturbances d

Set Constraint

ModelPredictiveControl.setconstraint! — Function

setconstraint!(mpc::PredictiveController; <keyword arguments>)

Set the constraint parameters of mpc predictive controller.

The predictive controllers support both soft and hard constraints, defined by:

\[\begin{alignat*}{3} \mathbf{u_{min} - c_{u_{min}}} ϵ &≤ \mathbf{u}(k+j) &&≤ \mathbf{u_{max} + c_{u_{max}}} ϵ &&\qquad j = 0, 1 ,..., H_c - 1 \\ \mathbf{Δu_{min} - c_{Δu_{min}}} ϵ &≤ \mathbf{Δu}(k+j) &&≤ \mathbf{Δu_{max} + c_{Δu_{max}}} ϵ &&\qquad j = 0, 1 ,..., H_c - 1 \\ \mathbf{ŷ_{min} - c_{ŷ_{min}}} ϵ &≤ \mathbf{ŷ}(k+j) &&≤ \mathbf{ŷ_{max} + c_{ŷ_{max}}} ϵ &&\qquad j = 1, 2 ,..., H_p \\ \end{alignat*}\]

and also $ϵ ≥ 0$. All the constraint parameters are vector. Use ±Inf values when there is no bound. The constraint softness parameters $\mathbf{c}$, also called equal concern for relaxation, are non-negative values that specify the softness of the associated bound. Use 0.0 values for hard constraints. The predicted output constraints $\mathbf{ŷ_{min}}$ and $\mathbf{ŷ_{max}}$ are soft by default.

Arguments

Info

The default constraints are mentioned here for clarity but omitting a keyword argument will not re-assign to its default value (defaults are set at construction only).

umin=fill(-Inf,nu) : manipulated input lower bounds $\mathbf{u_{min}}$

umax=fill(+Inf,nu) : manipulated input upper bounds $\mathbf{u_{max}}$

Δumin=fill(-Inf,nu) : manipulated input increment lower bounds $\mathbf{Δu_{min}}$

Δumax=fill(+Inf,nu) : manipulated input increment upper bounds $\mathbf{Δu_{max}}$

ŷmin=fill(-Inf,ny) : predicted output lower bounds $\mathbf{ŷ_{min}}$

ŷmax=fill(+Inf,ny) : predicted output upper bounds $\mathbf{ŷ_{max}}$

c_umin=fill(0.0,nu) : umin softness weights $\mathbf{c_{u_{min}}}$

c_umax=fill(0.0,nu) : umax softness weights $\mathbf{c_{u_{max}}}$

c_Δumin=fill(0.0,nu) : Δumin softness weights $\mathbf{c_{Δu_{min}}}$

c_Δumax=fill(0.0,nu) : Δumax softness weights $\mathbf{c_{Δu_{max}}}$

c_ŷmin=fill(1.0,ny) : ŷmin softness weights $\mathbf{c_{ŷ_{min}}}$

c_ŷmax=fill(1.0,ny) : ŷmax softness weights $\mathbf{c_{ŷ_{max}}}$

Move Manipulated Input

ModelPredictiveControl.moveinput! — Function

moveinput!(
    mpc::PredictiveController, 
    ry, 
    d  = Float64[];
    R̂y = repeat(ry, mpc.Hp), 
    D̂  = repeat(d,  mpc.Hp), 
    ym = nothing
)

Compute the optimal manipulated input value u for the current control period.

Solve the optimization problem of mpc PredictiveController and return the results $\mathbf{u}(k)$. Following the receding horizon principle, the algorithm discards the optimal future manipulated inputs $\mathbf{u}(k+1), \mathbf{u}(k+2),$... The arguments ry and d are current output setpoints $\mathbf{r_y}(k)$ and measured disturbances $\mathbf{d}(k)$. The predicted output setpoint R̂y and measured disturbances D̂ are defined as:

\[ \mathbf{R̂_y} = \begin{bmatrix} \mathbf{r̂_y}(k+1) \\ \mathbf{r̂_y}(k+2) \\ \vdots \\ \mathbf{r̂_y}(k+H_p) \end{bmatrix} \qquad \text{and} \qquad \mathbf{D̂} = \begin{bmatrix} \mathbf{d̂}(k+1) \\ \mathbf{d̂}(k+2) \\ \vdots \\ \mathbf{d̂}(k+H_p) \end{bmatrix}\]

They are assumed constant in the future by default, that is $\mathbf{r̂_y}(k+j) = \mathbf{r_y}(k)$ and $\mathbf{d̂}(k+j) = \mathbf{d}(k)$ for $j=1$ to $H_p$. Current measured outputs ym (keyword argument) are only required if mpc.estim is a InternalModel.

See also LinMPC, NonLinMPC.

Examples

julia> mpc = LinMPC(LinModel(tf(5, [2, 1]), 3), Nwt=[0], Hp=1000, Hc=1);

julia> u = moveinput!(mpc, [5]); round.(u, digits=3)
1-element Vector{Float64}:
 1.0

Get Additional Information

ModelPredictiveControl.getinfo — Function

getinfo(mpc::PredictiveController)

Get additional information about mpc controller optimum.

Return the dictionary info with the additional information, and sol_summary, the optimizer solution summary that can be printed. The dictionary info has the following fields:

:ΔU : optimal manipulated input increments over Hc $(\mathbf{ΔU})$

:ϵ : optimal slack variable $(ϵ)$

:J : objective value optimum $(J)$

:U : optimal manipulated inputs over Hp $(\mathbf{U})$

:u : current optimal manipulated input $(\mathbf{u})$

:d : current measured disturbance $(\mathbf{d})$

:D̂ : predicted measured disturbances over Hp $(\mathbf{D̂})$

:ŷ : current estimated output $(\mathbf{ŷ})$

:Ŷ : predicted outputs over Hp $(\mathbf{Ŷ = Ŷ_d + Ŷ_s})$

:Ŷd : predicted deterministic output over Hp $(\mathbf{Ŷ_d})$

:Ŷs : predicted stochastic output over Hp $(\mathbf{Ŷ_s})$

:R̂y : predicted output setpoint over Hp $(\mathbf{R̂_y})$

:R̂u : predicted manipulated input setpoint over Hp $(\mathbf{R̂_u})$

Examples

julia> mpc = LinMPC(LinModel(tf(5, [2, 1]), 3), Nwt=[0], Hp=1000, Hc=1);

julia> u = moveinput!(mpc, [5]);

julia> info, sol_summary = getinfo(mpc); round.(info[:Ŷ][end], digits=2)
5.0

getinfo(mpc::NonLinMPC)

Invoke getinfo(::PredictiveController) and add :JE the economic optimum $J_E$.