State Estimator Design

This package includes many state estimators (or state observer), both for deterministic and stochastic systems.

The estimator are all implemented in the predictor form (a.k.a. observer form), that is, they all estimates at each discrete time $k$ the states of the next period $\mathbf{x̂}_k(k+1)$. This form comes in handy for control applications since the estimations come after the controller computations, without introducing any additional delays. In contrast, the filter form that estimates $\mathbf{x̂}_k(k)$ is sometimes slightly more accurate.

Info

All the state estimators support measured $\mathbf{y^m}$ and unmeasured $\mathbf{y^u}$ model outputs, where $\mathbf{y}$ refers to all of them.

Abstract Types

ModelPredictiveControl.StateEstimator — Type

Abstract supertype of all state estimators.

(estim::StateEstimator)(d=Float64[])

Functor allowing callable StateEstimator object as an alias for evaloutput.

Examples

julia> kf = KalmanFilter(setop!(LinModel(tf(3, [10, 1]), 2), yop=[20]));

julia> ŷ = kf() 
1-element Vector{Float64}:
 20.0

SteadyKalmanFilter

ModelPredictiveControl.SteadyKalmanFilter — Type

SteadyKalmanFilter(model::LinModel; <keyword arguments>)

Construct a SteadyKalmanFilter (asymptotic) based on the LinModel model.

The steady-state Kalman filter is based on the process model :

\[\begin{aligned} \mathbf{x}(k+1) &= \mathbf{Â x}(k) + \mathbf{B̂_u u}(k) + \mathbf{B̂_d d}(k) + \mathbf{w}(k) \\ \mathbf{y^m}(k) &= \mathbf{Ĉ^m x}(k) + \mathbf{D̂_d^m d}(k) + \mathbf{v}(k) \\ \mathbf{y^u}(k) &= \mathbf{Ĉ^u x}(k) + \mathbf{D̂_d^u d}(k) \end{aligned}\]

with sensor $\mathbf{v}(k)$ and process $\mathbf{w}(k)$ noises as uncorrelated zero mean white noise processes, with a respective covariance of $\mathbf{R̂}$ and $\mathbf{Q̂}$. The arguments are in standard deviations σ, i.e. same units than outputs and states. The matrices $\mathbf{Â, B̂_u, B̂_d, Ĉ, D̂_d}$ are model matrices augmented with the stochastic model, which is specified by the numbers of output integrator nint_ym. Likewise, the covariance matrices are augmented with $\mathbf{Q̂ = \text{diag}(Q, Q_{int})}$ and $\mathbf{R̂ = R}$. The matrices $\mathbf{Ĉ^m, D̂_d^m}$ are the rows of $\mathbf{Ĉ, D̂_d}$ that correspond to measured outputs $\mathbf{y^m}$(and unmeasured ones for $\mathbf{Ĉ^u, D̂_d^u}$).

Arguments

model::LinModel : (deterministic) model for the estimations.

i_ym=1:model.ny : model output indices that are measured $\mathbf{y^m}$, the rest are unmeasured $\mathbf{y^u}$.

σQ=fill(0.1,model.nx) : main diagonal of the process noise covariance $\mathbf{Q}$ of model, specified as a standard deviation vector.

σR=fill(0.1,length(i_ym)) : main diagonal of the sensor noise covariance $\mathbf{R}$ of model measured outputs, specified as a standard deviation vector.

nint_ym=fill(1,length(i_ym)) : integrator quantity per measured outputs (vector) for the stochastic model, use nint_ym=0 for no integrator at all.

σQ_int=fill(0.1,sum(nint_ym)) : same than σQ but for the stochastic model covariance $\mathbf{Q_{int}}$ (composed of output integrators)

Examples

julia> model = LinModel([tf(3, [30, 1]); tf(-2, [5, 1])], 0.5);

julia> estim = SteadyKalmanFilter(model, i_ym=[2], σR=[1], σQ_int=[0.01])
SteadyKalmanFilter state estimator with a sample time Ts = 0.5 s and:
 1 manipulated inputs u
 3 states x̂
 1 measured outputs ym
 1 unmeasured outputs yu
 0 measured disturbances d

Extended Help

The model augmentation with nint_ym vector produces the integral action when the estimator is used in a controller as state feedback (a.k.a. offset-free control). The default is 1 integrator per measured outputs. More than 1 integrator is interesting only when model is integrating or unstable, or when the unmeasured output disturbances are "ramp-like". See augment_model.

Tip

Increasing σQ_int values increases the integral action "gain".

KalmanFilter

ModelPredictiveControl.KalmanFilter — Type

KalmanFilter(model::LinModel; <keyword arguments>)

Construct a time-varying KalmanFilter based on the LinModel model.

The process model is identical to SteadyKalmanFilter.

Arguments

model::LinModel : (deterministic) model for the estimations.

σP0=fill(10,model.nx) : main diagonal of the initial estimate covariance $\mathbf{P}(0)$, specified as a standard deviation vector.

σP0_int=fill(10,sum(nint_ym)) : same than σP0 but for the stochastic model covariance $\mathbf{P_{int}}(0)$ (composed of output integrators).

<keyword arguments> of SteadyKalmanFilter

Examples

julia> model = LinModel([tf(3, [30, 1]); tf(-2, [5, 1])], 0.5);

julia> estim = KalmanFilter(model, i_ym=[2], σR=[1], σP0=[100, 100], σQ_int=[0.01])
KalmanFilter state estimator with a sample time Ts = 0.5 s and:
 1 manipulated inputs u
 3 states x̂
 1 measured outputs ym
 1 unmeasured outputs yu
 0 measured disturbances d

InternalModel

ModelPredictiveControl.InternalModel — Type

InternalModel(model::SimModel; i_ym=1:model.ny, stoch_ym=ss(1,1,1,1,model.Ts).*I)

Construct an InternalModel estimator based on model (LinModel or NonLinModel).

i_ym provides the model output indices that are measured $\mathbf{y^m}$, the rest are unmeasured $\mathbf{y^u}$. model evaluates the deterministic predictions $\mathbf{ŷ_d}$, and stoch_ym, the stochastic predictions of the measured outputs $\mathbf{ŷ_s^m}$ (the unmeasured ones being $\mathbf{ŷ_s^u=0}$). The predicted outputs sum both values : $\mathbf{ŷ = ŷ_d + ŷ_s}$.

Warning

InternalModel estimator does not work if model is integrating or unstable. The constructor verifies these aspects for LinModel but not for NonLinModel. Uses any other state estimator in such cases.

Examples

julia> estim = InternalModel(LinModel([tf(3, [30, 1]); tf(-2, [5, 1])], 0.5), i_ym=[2])
InternalModel state estimator with a sample time Ts = 0.5 s and:
 1 manipulated inputs u
 2 states x̂
 1 measured outputs ym
 1 unmeasured outputs yu
 0 measured disturbances d

Extended Help

stoch_ym is a TransferFunction or StateSpace model that hypothetically filters a zero mean white noise vector. Its default value supposes 1 integrator per measured outputs, assuming that the current stochastic estimate $\mathbf{ŷ_s^m}(k) = \mathbf{y^m}(k) - \mathbf{ŷ_d^m}(k)$ will be constant in the future. This is the dynamic matrix control (DMC) strategy, which is simple but sometimes too aggressive. Additional poles and zeros in stoch_ym can mitigate this.

Luenberger

UnscentedKalmanFilter

MovingHorizonEstimator

Generic Functions

ModelPredictiveControl.initstate! — Function

initstate!(estim::StateEstimator, u, ym, d=Float64[])

Init estim.x̂ states from current inputs u, measured outputs ym and disturbances d.

The method tries to find a good steady-state to initialize estim.x̂ estimate :

If estim.model is a non-integrating LinModel, it evaluates estim.model steady-state with current inputs u and measured disturbances d, and saves the result to estim.x̂[1:nx].

Else, the current deterministic states estim.x̂[1:nx] are left unchanged (use setstate! to manually modify them).

It then estimates the measured outputs ŷm from these states, and the residual offset with current measured outputs (ym - ŷm) initializes the integrators of the stochastic model. This approach ensures that $\mathbf{ŷ^m}(0) = \mathbf{y^m}(0)$. For LinModel, it also ensures that the estimator starts at steady-state, resulting in a bumpless manual to automatic transfer for control applications.

Examples

julia> estim = SteadyKalmanFilter(LinModel(tf(3, [10, 1]), 0.5), nint_ym=[2]);

julia> x̂ = initstate!(estim, [1], [3 - 0.1])
3-element Vector{Float64}:
  5.0000000000000115
  0.0
 -0.10000000000000675

initstate!(estim::InternalModel, u, ym, d=Float64[])

Init estim.x̂d / x̂s states from current inputs u, meas. outputs ym and disturb. d.

The deterministic state estim.x̂d initialization method is identical to initstate!(::StateEstimator). The stochastic states estim.x̂s are init at 0.

initstate!(estim::KalmanFilter, u, ym, d=Float64[])

Initialize covariance estim.P̂ and invoke initstate!(::StateEstimator).

initstate!(mpc::PredictiveController, u, ym, d=Float64[])

Init mpc controller variables and the states of mpc.estim StateEstimator.