State estimator design

Info

All the state estimators support measured $\mathbf{y^m}$ and unmeasured $\mathbf{y^u}$ model outputs, where $\mathbf{y}$ refers to all of them.

StateEstimator functions and types

ModelPredictiveControl.StateEstimator — Type

Abstract supertype of all state estimators.

(estim::StateEstimator)(d=Float64[])

Functor allowing callable StateEstimator object as an alias for evaloutput.

Examples

julia> kf = KalmanFilter(setop!(LinModel(tf(3, [10, 1]), 2), yop=[20]));

julia> ŷ = kf() 
1-element Vector{Float64}:
 20.0

InternalModel

ModelPredictiveControl.InternalModel — Type

InternalModel(model::SimModel; i_ym=1:model.ny, stoch_ym=ss(1,1,1,1,model.Ts).*I)

Construct an InternalModel estimator based on model.

i_ym provides the model output indices that are measured $\mathbf{y^m}$, the rest are unmeasured $\mathbf{y^u}$. model evaluates the deterministic predictions $\mathbf{ŷ_d}$, and stoch_ym, the stochastic predictions of the measured outputs $\mathbf{ŷ_s^m}$ (the unmeasured ones being $\mathbf{ŷ_s^u} = \mathbf{0}$).

stoch_ym is a TransferFunction or StateSpace model that hypothetically filters a zero mean white noise vector. Its default value supposes 1 integrator per measured outputs, assuming that the current stochastic estimate $\mathbf{ŷ_s^m}(k) = \mathbf{y^m}(k) - \mathbf{ŷ_d^m}(k)$ will be constant in the future. This is the dynamic matrix control (DMC) strategy, which is simple but sometimes too aggressive. Additional poles and zeros in stoch_ym can mitigate this.

Warning

InternalModel estimator does not work if model is integrating or unstable. The constructor verifies these aspects for LinModel but not for NonLinModel. Uses any other state estimator in such cases.

See also LinModel, NonLinModel

Examples

julia> estim = InternalModel(LinModel([tf(3, [30, 1]); tf(-2, [5, 1])], 0.5), i_ym=[2])
InternalModel state estimator with a sample time Ts = 0.5 s and:
 1 manipulated inputs u
 2 states x̂
 1 measured outputs ym
 1 unmeasured outputs yu
 0 measured disturbances d

(estim::InternalModel)(ym, d=Float64[])

Functor allowing callable InternalModel object as an alias for evaloutput.

ModelPredictiveControl.updatestate! — Method

updatestate!(estim::InternalModel, u, ym, d=Float64[])

Update estim.x̂dx̂s states with current inputs u, measured outputs ym and dist. d.

ModelPredictiveControl.evaloutput — Method

evaloutput(estim::InternalModel, ym, d=Float64[])

Evaluate InternalModel outputs ̂ŷ from estim.x̂d states.

ym and d are current measured outputs and disturbances, respectively. InternalModel estimator needs current measured outputs $\mathbf{y^m}(k)$ to estimate its outputs $\mathbf{ŷ}(k)$, since the strategy imposes that $\mathbf{ŷ^m}(k) = \mathbf{y^m}(k)$ is always true.

Luenburger

SteadyKalmanFilter

KalmanFilter

ModelPredictiveControl.KalmanFilter — Type

KalmanFilter(model::LinModel; <keyword arguments>)

Construct a time-varying KalmanFilter based on model.

The process model is :

Arguments

model::LinModel : (deterministic) model for the estimations.

i_ym=1:model.ny : model output indices that are measured $\mathbf{y^m}$, the rest are unmeasured $\mathbf{y^u}$.

σP0=fill(10,model.nx) : standard deviation vector for the initial estimate covariance $\mathbf{P}(0)$ of model.

σQ=fill(0.1,model.nx) : standard deviation vector for the process noise covariance $\mathbf{Q}$ of model.

σR=fill(0.1,length(i_ym)) : standard deviation vector for the sensor noise covariance $\mathbf{R}$ of model measured outputs.

nint_ym=fill(1,length(i_ym)) : integrator quantity per measured outputs for the stochastic model, nint_ym=0 means no integrator at all.

σP0_int=fill(10,sum(nint_ym)) : standard deviation vector for the initial estimate covariance of the stochastic model (composed of output integrators).

σQ_int=fill(0.1,sum(nint_ym)) : standard deviation vector for the process noise covariance of the stochastic model (composed of output integrators).

ModelPredictiveControl.updatestate! — Method

updatestate!(estim::KalmanFilter, u, ym, d=Float64[])

Update estim.x̂P̂ estimates with current inputs u, measured outputs ym and dist. d.

ModelPredictiveControl.evaloutput — Method

evaloutput(estim::KalmanFilter, d=Float64[])

Evaluate KalmanFilter outputs ̂ŷ from estim.x̂ states and current disturbances d.

UnscentedKalmanFilter

MovingHorizonEstimator

Advanced Topics

Internals

ModelPredictiveControl.init_internalmodel — Function

init_internalmodel(As, Bs, Cs, Ds)

Calc stochastic model update matrices Âs and B̂s for InternalModel estimator.

As, Bs, Cs and Ds are the stochastic model matrices :

\[\begin{aligned} \mathbf{x_s}(k+1) &= \mathbf{A_s x_s}(k) + \mathbf{B_s e}(k) \\ \mathbf{y_s}(k) &= \mathbf{C_s x_s}(k) + \mathbf{D_s e}(k) \end{aligned}\]

where $\mathbf{e}(k)$ is conceptual and unknown zero mean white noise. Its optimal estimation is $\mathbf{ê=0}$, the expected value. Thus, the Âs and B̂s matrices that optimally update the stochastic estimate $\mathbf{x̂_s}$ are:

\[\begin{aligned} \mathbf{x̂_s}(k+1) &= \mathbf{(A_s - B_s D_s^{-1} C_s) x̂_s}(k) + \mathbf{(B_s D_s^{-1}) ŷ_s}(k) \\ &= \mathbf{Â_s x̂_s}(k) + \mathbf{B̂_s ŷ_s}(k) \end{aligned}\]

with current stochastic outputs estimation $\mathbf{ŷ_s}(k)$, composed of the measured $\mathbf{ŷ_s^m}(k) = \mathbf{y^m}(k) - \mathbf{ŷ_d^m}(k)$ and unmeasured $\mathbf{ŷ_s^u = 0}$ outputs. See [1].

ModelPredictiveControl.init_estimstoch — Function

Asm, Csm = init_estimstoch(model::SimModel, i_ym, nint_ym)

Calc stochastic model matrices from output integrators specifications for state estimation.

For closed-loop state estimators. nint_ym is a vector providing how many integrator should be added for each measured output $\mathbf{y^m}$. The argument generates the Asm and Csm matrices:

\[\begin{aligned} \mathbf{x_s}(k+1) &= \mathbf{A_s^m x_s}(k) + \mathbf{B_s^m e}(k) \\ \mathbf{y_s^m}(k) &= \mathbf{C_s^m x_s}(k) \end{aligned}\]

where $\mathbf{e}(k)$ is conceptual and unknown zero mean white noise. $\mathbf{B_s^m}$ is not used for closed-loop state estimators thus ignored.

ModelPredictiveControl.augment_model — Function

augment_model(model::LinModel, As, Cs)

Augment LinModel state-space matrices with the stochastic ones As and Cs.

If $\mathbf{x_d}$ are model.x states, and $\mathbf{x_s}$, the states defined at init_estimstoch, we define an augmented state vector $\mathbf{x} = [ \begin{smallmatrix} \mathbf{x_d} \\ \mathbf{x_s} \end{smallmatrix} ]$. The function returns the augmented model matrices Â, B̂u, Ĉ, B̂d and D̂d:

\[\begin{aligned} \mathbf{x}(k+1) &= \mathbf{Â x}(k) + \mathbf{B̂_u u}(k) + \mathbf{B̂_d d}(k) \\ \mathbf{y}(k) &= \mathbf{Ĉ x}(k) + \mathbf{D̂_d d}(k) \end{aligned}\]

1Desbiens, A., D. Hodouin & É. Plamondon. 2000, "Global predictive control : a unified control structure for decoupling setpoint tracking, feedforward compensation and disturbance rejection dynamics", IEE Proceedings - Control Theory and Applications, vol. 147, no 4, https://doi.org/10.1049/ip-cta:20000443, p. 465–475, ISSN 1350-2379.