State estimator design

This package includes many state estimators (or state observer), both for deterministic and stochastic systems.

The estimator are all implemented in the predictor form (a.k.a. observer form), that is, they all estimates at each discrete time $k$ the states of the next period $\mathbf{x̂}_k(k+1)$. This form comes in handy for control applications since the estimations come after the controller computations, without introducing any additional delays. In contrast, the filter form that estimates $\mathbf{x̂}_k(k)$ is sometimes slightly more accurate.

Info

All the state estimators support measured $\mathbf{y^m}$ and unmeasured $\mathbf{y^u}$ model outputs, where $\mathbf{y}$ refers to all of them.

StateEstimator types

ModelPredictiveControl.StateEstimator — Type

Abstract supertype of all state estimators.

(estim::StateEstimator)(d=Float64[])

Functor allowing callable StateEstimator object as an alias for evaloutput.

Examples

julia> kf = KalmanFilter(setop!(LinModel(tf(3, [10, 1]), 2), yop=[20]));

julia> ŷ = kf() 
1-element Vector{Float64}:
 20.0

InternalModel

ModelPredictiveControl.InternalModel — Type

InternalModel(model::SimModel; i_ym=1:model.ny, stoch_ym=ss(1,1,1,1,model.Ts).*I)

Construct an InternalModel estimator based on model (LinModel or NonLinModel).

i_ym provides the model output indices that are measured $\mathbf{y^m}$, the rest are unmeasured $\mathbf{y^u}$. model evaluates the deterministic predictions $\mathbf{ŷ_d}$, and stoch_ym, the stochastic predictions of the measured outputs $\mathbf{ŷ_s^m}$ (the unmeasured ones being $\mathbf{ŷ_s^u=0}$). The predicted outputs sum both values : $\mathbf{ŷ = ŷ_d + ŷ_s}$.

Warning

InternalModel estimator does not work if model is integrating or unstable. The constructor verifies these aspects for LinModel but not for NonLinModel. Uses any other state estimator in such cases.

See also LinModel, NonLinModel

Examples

julia> estim = InternalModel(LinModel([tf(3, [30, 1]); tf(-2, [5, 1])], 0.5), i_ym=[2])
InternalModel state estimator with a sample time Ts = 0.5 s and:
 1 manipulated inputs u
 2 states x̂
 1 measured outputs ym
 1 unmeasured outputs yu
 0 measured disturbances d

Extended Help

stoch_ym is a TransferFunction or StateSpace model that hypothetically filters a zero mean white noise vector. Its default value supposes 1 integrator per measured outputs, assuming that the current stochastic estimate $\mathbf{ŷ_s^m}(k) = \mathbf{y^m}(k) - \mathbf{ŷ_d^m}(k)$ will be constant in the future. This is the dynamic matrix control (DMC) strategy, which is simple but sometimes too aggressive. Additional poles and zeros in stoch_ym can mitigate this.

Luenberger

SteadyKalmanFilter

ModelPredictiveControl.SteadyKalmanFilter — Type

SteadyKalmanFilter(model::LinModel; <keyword arguments>)

Construct a SteadyKalmanFilter (steady-state or asymptotic) based on model.

The process model is the stochastic system :

\[\begin{aligned} \mathbf{x}(k+1) &= \mathbf{Â x}(k) + \mathbf{B̂_u u}(k) + \mathbf{B̂_d d}(k) + \mathbf{w}(k) \\ \mathbf{y^m}(k) &= \mathbf{Ĉ^m x}(k) + \mathbf{D̂_d^m d}(k) + \mathbf{v}(k) \\ \mathbf{y^u}(k) &= \mathbf{Ĉ^u x}(k) + \mathbf{D̂_d^u d}(k) \end{aligned}\]

with sensor $\mathbf{v}(k)$ and process $\mathbf{w}(k)$ noises as zero mean white noise processes, with a respective covariance of $\mathbf{R̂}$ and $\mathbf{Q̂}$. The arguments are in standard deviations σ, i.e. same units than outputs and states. The matrices $\mathbf{Â, B̂_u, B̂_d, Ĉ, D̂_d}$ are model matrices augmented with the stochastic model, which is specified by the numbers of output integrator nint_ym. Likewise, the covariance matrices are augmented with $\mathbf{Q̂ = \text{diag}(Q, Q_{int})}$ and $\mathbf{R̂ = R}$. The matrices $\mathbf{Ĉ^m, D̂_d^m}$ are the rows of $\mathbf{Ĉ, D̂_d}$ that correspond to measured outputs $\mathbf{y^m}$ (and unmeasured ones for $\mathbf{Ĉ^u, D̂_d^u}$).

See also LinModel.

Arguments

model::LinModel : (deterministic) model for the estimations.

i_ym=1:model.ny : model output indices that are measured $\mathbf{y^m}$, the rest are unmeasured $\mathbf{y^u}$.

σQ=fill(0.1,model.nx) : main diagonal of the process noise covariance $\mathbf{Q}$ of model, specified as a standard deviation vector.

σR=fill(0.1,length(i_ym)) : main diagonal of the sensor noise covariance $\mathbf{R}$ of model measured outputs, specified as a standard deviation vector.

nint_ym=fill(1,length(i_ym)) : integrator quantity per measured outputs (vector) for the stochastic model, use nint_ym=0 for no integrator at all.

σQ_int=fill(0.1,sum(nint_ym)) : same than σQ but for the stochastic model covariance $\mathbf{Q_{int}}$ (composed of output integrators)

Extended Help

The model augmentation with nint_ym vector produces the integral action when the estimator is used in a controller as state feedback (a.k.a. offset-free control). The default is 1 integrator per measured outputs. More than 1 integrator is interesting only when model is integrating or unstable, or when the unmeasured output disturbances are "ramp-like". See augment_model.

Tip

Increasing σQ_int values increases the integral action "gain".

KalmanFilter

ModelPredictiveControl.KalmanFilter — Type

KalmanFilter(model::LinModel; <keyword arguments>)

Construct a time-varying KalmanFilter based on model.

The process model is identical to SteadyKalmanFilter.

Arguments

model::LinModel : (deterministic) model for the estimations.

σP0=fill(10,model.nx) : main diagonal of the initial estimate covariance $\mathbf{P}(0)$, specified as a standard deviation vector.

σP0_int=fill(10,sum(nint_ym)) : same than σP0 but for the stochastic model covariance $\mathbf{P_{int}}(0)$ (composed of output integrators).

<keyword arguments> of SteadyKalmanFilter

UnscentedKalmanFilter

MovingHorizonEstimator

StateEstimator functions

ModelPredictiveControl.initstate! — Function

initstate!(estim::StateEstimator, u, ym, d=Float64[])

Init estim.x̂ states from current inputs u, measured outputs ym and disturbances d.

The method tries to find a good steady-state to initialize estim.x̂ estimate :

If estim.model is a LinModel, it evaluates estim.model steady-state with current manipulated inputs u and measured disturbances d.

If estim.model is a NonLinModel, the current deterministic states estim.x̂[1:nx] are left unchanged (use setstate! to manually modify them).

It then estimates the measured outputs ŷm from these states, and the residual offset with current measured outputs (ym - ŷm) initializes the integrators of the stochastic model. This approach ensures that $\mathbf{ŷ^m}(0) = \mathbf{y^m}(0)$. For LinModel, it also ensures that the estimator starts at steady-state, resulting in a bumpless manual to automatic transfer for control applications.

initstate!(estim::InternalModel, u, ym, d=Float64[])

Init estim.x̂d / x̂s states from current inputs u, meas. outputs ym and disturb. d.

The deterministic state estim.x̂d initialization method is identical to initstate!(::StateEstimator). The stochastic states estim.x̂s are init at 0.

initstate!(estim::KalmanFilter, u, ym, d=Float64[])

Initialize covariance estim.P̂ and invoke initstate!(::StateEstimator).

ModelPredictiveControl.updatestate! — Function

updatestate!(model::SimModel, u, d=Float64[])

Update model.x states with current inputs u and measured disturbances d.

updatestate!(estim::InternalModel, u, ym, d=Float64[])

Update estim.x̂ \ x̂d \ x̂s with current inputs u, measured outputs ym and dist. d.

updatestate!(estim::SteadyKalmanFilter, u, ym, d=Float64[])

Update estim.x̂ with current inputs u, measured outputs ym and dist. d.

updatestate!(estim::KalmanFilter, u, ym, d=Float64[])

Update estim.x̂ \ P̂ with current inputs u, measured outputs ym and dist. d.

ModelPredictiveControl.evaloutput — Function

evaloutput(model::SimModel, d=Float64[])

Evaluate SimModel outputs y from model.x states and measured disturbances d.

evaloutput(estim::InternalModel, ym, d=Float64[])

Evaluate InternalModel outputs ̂ŷ from estim.x̂d states.

ym and d are current measured outputs and disturbances, respectively. InternalModel estimator needs current measured outputs $\mathbf{y^m}(k)$ to estimate its outputs $\mathbf{ŷ}(k)$, since the strategy imposes that $\mathbf{ŷ^m}(k) = \mathbf{y^m}(k)$ is always true.

evaloutput(estim::Union{SteadyKalmanFilter, KalmanFilter}, d=Float64[])

Evaluate Kalman filter outputs ̂ŷ from estim.x̂ states and current disturbances d.

ModelPredictiveControl.setstate! — Function

setstate!(model::SimModel, x)

Set model.x states to values specified by x.

setstate!(estim::StateEstimator, x̂)

Set estim.x̂ states to values specified by x̂.

Advanced Topics

Internals

ModelPredictiveControl.init_internalmodel — Function

init_internalmodel(As, Bs, Cs, Ds)

Calc stochastic model update matrices Âs and B̂s for InternalModel estimator.

As, Bs, Cs and Ds are the stochastic model matrices :

\[\begin{aligned} \mathbf{x_s}(k+1) &= \mathbf{A_s x_s}(k) + \mathbf{B_s e}(k) \\ \mathbf{y_s}(k) &= \mathbf{C_s x_s}(k) + \mathbf{D_s e}(k) \end{aligned}\]

where $\mathbf{e}(k)$ is conceptual and unknown zero mean white noise. Its optimal estimation is $\mathbf{ê=0}$, the expected value. Thus, the Âs and B̂s matrices that optimally update the stochastic estimate $\mathbf{x̂_s}$ are:

\[\begin{aligned} \mathbf{x̂_s}(k+1) &= \mathbf{(A_s - B_s D_s^{-1} C_s) x̂_s}(k) + \mathbf{(B_s D_s^{-1}) ŷ_s}(k) \\ &= \mathbf{Â_s x̂_s}(k) + \mathbf{B̂_s ŷ_s}(k) \end{aligned}\]

with current stochastic outputs estimation $\mathbf{ŷ_s}(k)$, composed of the measured $\mathbf{ŷ_s^m}(k) = \mathbf{y^m}(k) - \mathbf{ŷ_d^m}(k)$ and unmeasured $\mathbf{ŷ_s^u = 0}$ outputs. See [1].

ModelPredictiveControl.init_estimstoch — Function

Asm, Csm = init_estimstoch(model::SimModel, i_ym, nint_ym::Vector{Int})

Calc stochastic model matrices from output integrators specifications for state estimation.

For closed-loop state estimators. nint_ym is a vector providing how many integrator should be added for each measured output $\mathbf{y^m}$. The argument generates the Asm and Csm matrices:

\[\begin{aligned} \mathbf{x_s}(k+1) &= \mathbf{A_s^m x_s}(k) + \mathbf{B_s^m e}(k) \\ \mathbf{y_s^m}(k) &= \mathbf{C_s^m x_s}(k) \end{aligned}\]

where $\mathbf{e}(k)$ is conceptual and unknown zero mean white noise. $\mathbf{B_s^m}$ is not used for closed-loop state estimators thus ignored.

ModelPredictiveControl.augment_model — Function

augment_model(model::LinModel, As, Cs)

Augment LinModel state-space matrices with the stochastic ones As and Cs.

If $\mathbf{x_d}$ are model.x states, and $\mathbf{x_s}$, the states defined at init_estimstoch, we define an augmented state vector $\mathbf{x} = [ \begin{smallmatrix} \mathbf{x_d} \\ \mathbf{x_s} \end{smallmatrix} ]$. The function returns the augmented model matrices Â, B̂u, Ĉ, B̂d and D̂d:

\[\begin{aligned} \mathbf{x}(k+1) &= \mathbf{Â x}(k) + \mathbf{B̂_u u}(k) + \mathbf{B̂_d d}(k) \\ \mathbf{y}(k) &= \mathbf{Ĉ x}(k) + \mathbf{D̂_d d}(k) \end{aligned}\]

1Desbiens, A., D. Hodouin & É. Plamondon. 2000, "Global predictive control : a unified control structure for decoupling setpoint tracking, feedforward compensation and disturbance rejection dynamics", IEE Proceedings - Control Theory and Applications, vol. 147, no 4, https://doi.org/10.1049/ip-cta:20000443, p. 465–475, ISSN 1350-2379.